% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_comm.R
\name{active_connection}
\alias{active_connection}
\alias{[DBI::dbIsValid]}
\title{Is a connection object still available?}
\usage{
active_connection(db_conn = con)
}
\arguments{
\item{db_conn}{connection object (default "con")}
}
\value{
LGL scalar indicating whether the database is available
}
\description{
This is a thin wrapper for [DBI::dbIsValid] with some error logging.
}
