% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{add_help}
\alias{add_help}
\title{Attach a superscript icon with a bsTooltip to an HTML element}
\usage{
add_help(
  id,
  tooltip,
  icon_name = "question",
  size = "xs",
  icon_class = "info-tooltip primary",
  ...
)
}
\arguments{
\item{id}{CHR scalar of the HTML ID to which to append the icon}

\item{tooltip}{CHR scalar of the tooltip text}

\item{icon_name}{CHR scalar of the icon name, which must be understandable by
the `shiny::icon` function; e.g. a font-awesome icon name (default:
"question").}

\item{size}{CHR scalar of the general icon size as understandable by the
font-awesome library (default: "xs")}

\item{icon_class}{CHR vector of classes to apply to the `<sup>` container,
as defined in the current CSS (default: "info-tooltip primary")}

\item{...}{Other named arguments to be passed to `shinyBS:bsTooltip`}
}
\value{
LIST of HTML tags for the desired help icon and its tooltip
}
\description{
Attach a superscript icon with a bsTooltip to an HTML element
}
\note{
The following CSS is typically defined to go with this.
 .info-tooltip {
   opacity: 30%;
   transition: opacity .25s;
 }
 .info-tooltip:hover {
   opacity: 100%;
 }
 .primary {
   color: #3c8dbc;
 }
}
\examples{
add_help("example", "a tooltip")

}
