% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_comm.R
\name{add_normalization_value}
\alias{add_normalization_value}
\title{Add value(s) to a normalization table}
\usage{
add_normalization_value("norm_table", name = "new value", acronym = "NV")
}
\arguments{
\item{db_table}{CHR scalar of the normalization table's name}

\item{db_conn}{connection object (default "con")}

\item{log_ns}{CHR scalar of the logging namespace to use during execution
(default: "db")}

\item{id_column}{CHR scalar of the column to use as the primary key
identifier for `db_table` (default: "id")}

\item{database_map}{LIST of the database entity relationship map, typically
from calling [er_map]. If NULL (default) the object "db_map" will be
searched for and used by default, otherwise it will be created with
[er_map]}

\item{...}{CHR vector of additional named arguments to be added; names not
appearing in the referenced table will be ignored}
}
\value{
NULL if unable to add the values, INT scalar of the new ID otherwise
}
\description{
One of the most common database operations is to look up or add a value in a
normalization table. This utility function adds a single value and returns
its associated id by using [build_db_action]. This is only suitable for a
single value. If you need to bulk add multiple new values, use this with
something like [lapply].
}
