% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementalcomposition.R
\name{adduct_formula}
\alias{adduct_formula}
\title{Add Adduct to Formula}
\usage{
adduct_formula(elementalformula, adduct = "+H")
}
\arguments{
\item{elementalformula}{character string elemental formula}

\item{adduct}{character string adduct state to add to the elemental formula, must contain an element, options are `+H`, `-H`, `+Na`, `+K`}
}
\value{
character string containing elemental formula with adduct
}
\description{
Add Adduct to Formula
}
\examples{
adduct_formula("C2H5O", adduct = "+H")
}
