% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_control.R
\name{api_start}
\alias{api_start}
\title{Start the plumber API}
\usage{
api_start(plumber_file = NULL, on_host = NULL, on_port = NULL)
}
\arguments{
\item{plumber_file}{CHR scalar of the path to a plumber API to launch
(default: NULL)}

\item{on_host}{CHR scalar of the host IP address (default: NULL)}

\item{on_port}{CHR or INT scalar of the host port to use (default: NULL)}
}
\value{
LGL scalar with success status
}
\description{
This is a wrapper to [plumber::pr_run] pointing to a project's opinionated
plumber settings with some error trapping. The host, port, and plumber file
are set in the "config/env_R.R" location as PLUMBER_HOST, PLUMBER_PORT, and
PLUMBER_FILE respectively.
}
\note{
If either of `on_host` or `on_port` are NULL they will default first to
  any existing environment values of PLUMBER_HOST and PLUMBER_PORT, then to
  getOption("plumber.host", "127.0.0.1") and getOption("plumber.port", 8080)

This will fail if the requested port is in use.
}
