% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_control.R
\name{api_stop}
\alias{api_stop}
\title{Stop the plumber API}
\usage{
api_stop(pr = NULL, flush = TRUE, db_conn = "con", remove_service_obj = TRUE)
}
\arguments{
\item{pr}{CHR scalar name of the plumber service object, typically only
created as a background observer from [callr::r_bg] as a result of calling
[api_reload] (default: NULL gets the environment setting for
PLUMBER_OBJ_NAME)}

\item{flush}{LGL scalar of whether to disconnect and reconnect to a database
connection named as `db_conn` (default: TRUE)}

\item{db_conn}{CHR scalar of the connection object name (default: "con")}

\item{remove_service_obj}{LGL scalar of whether to remove the reference to
`pr` from the current global environment (default: TRUE)}
}
\value{
None, stops the plumber server
}
\description{
Stop the plumber API
}
\note{
This will also kill and restart the connection object if `flush` is
  TRUE to release connections with certain configurations such as SQLite in
  write ahead log mode.

This function assumes the object referenced by name `pr` exists in the
  global environment, and `remove_service_object` will only remove it from
  .GlobalEnv.
}
