% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{append_icon_to}
\alias{append_icon_to}
\title{Create the JS to append an icon to an HTML element by its ID}
\usage{
append_icon_to(id, icon_name, icon_class = NULL)
}
\arguments{
\item{id}{CHR scalar of the HTML ID to which to append an icon}

\item{icon_name}{CHR scalar of the icon name, which must be understandable by
the `shiny::icon` function; e.g. a font-awesome icon name.}

\item{icon_class}{CHR vector of classes to apply}
}
\value{
CHR scalar suitable to execute with `shinyjs::runJS`
}
\description{
Create the JS to append an icon to an HTML element by its ID
}
\examples{
append_icon_to("example", "r-project", "fa-3x")

}
