% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_comm.R
\name{build_db}
\alias{build_db}
\title{Build or rebuild the database from scratch}
\usage{
build_db(db = "test_db.sqlite", db_conn_name = "test_conn")
}
\arguments{
\item{db}{CHR scalar of the database name (default: session value DB_NAME)}

\item{build_from}{CHR scalar of a SQL build script to use (default:
environment value DB_BUILD_FILE)}

\item{populate}{LGL scalar of whether to populate with data from the file in
`populate_with` (default: TRUE)}

\item{populate_with}{CHR scalar for the populate script (e.g.
"populate_demo.sql") to during after the build is complete; (default:
session value DB_DATA); ignored if `populate = FALSE`}

\item{archive}{LGL scalar of whether to create an archive of the current
database (if it exists) matching the name supplied in argument `db`
(default: FALSE), passed to [`remove_db()`]}

\item{sqlite_cli}{CHR scalar to use to look for installed sqlite3 CLI tools
in the current system environment (default: session value SQLITE_CLI)}

\item{connect}{LGL scalar of whether or not to connect to the rebuilt
database in the global environment as object `con`` (default: FALSE)}
}
\value{
None, check console for details
}
\description{
This function will build or rebuild the NIST HRAMS database structure from
scratch, removing the existing instance. By default, most parameters are set
in the environment (at "./config/env_glob.txt") but any values can be passed
directly. This can be used to quickly spin up multiple copies with a clean
slate using different build files, data files, or return to the last stable
release.
}
\details{
If sqlite3 and its command line interface are available on your platform,
that will be used (preferred method) but, if not, this function will read in
all the necessary files to directly create it using shell commands. The shell
method may not be universally applicable to certain compute environments or
may require elevated permissions.
}
