% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_sqlite_scripts.R
\name{build_triggers}
\alias{build_triggers}
\title{Build pairs of INSERT/UPDATE triggers to resolve foreign key relationships}
\usage{
build_triggers(er_map(db_conn = con))
}
\arguments{
\item{db_map}{LIST object containing descriptions of table mapping in an
opinionated manner, generally generated by [er_map]. The expectation is a
list of tables, with references in SQL form enumerated in a child element
with a name matching `references_in`}

\item{references_in}{CHR scalar naming the child element containing SQL
references statements of the form "fk_column REFERENCES table(pk_column)"
(default: "references" is provided by [er_map])}

\item{create_insert_trigger}{LGL scalar indicating whether to build an insert
trigger for each table (default: TRUE).}

\item{create_update_trigger}{LGL scalar indicating whether to build an update
trigger for each table (default: FALSE).}

\item{save_to_file}{CHR scalar of a file in which to write the output, if any
(default: NULL will return the resulting object to the R session)}
}
\value{
LIST object containing one element for each table in `db_map`
  containing foreign key references, with one child
}
\description{
When building schema by script, it is often handy to enforce certain
behaviors on database transactions involving foreign keys, especially in
SQLite. Given a properly structured list object describing the mappings
between tables in a schema (e.g. one deriving from [er_map]), this function
will parse those for foreign key relationships.
}
\details{
Primarily, this requires a list object referring to tables that contains in
each element a child element with the name provided in `references_in`. The
pre-pass parsing function [get_fkpk_relationships] is used to pull references
from
the full map is used.
}
\note{
Tables in `db_map` that do not contain foreign key relationships will
  be dropped from the output list.

This is largely a convenience function to programmatically apply
  [make_sql_triggers] to an entire schema. To skip tables with defined
  foreign key relationships for which triggers are undesirable, remove those
  tables from `db_map` prior to calling this function.
}
