% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_sqlite_scripts.R
\name{build_views}
\alias{build_views}
\title{Build SQL to create views on normalized tables in SQLite}
\usage{
build_views(db_map = er_map(con), dictionary = data_dictionary(con))
}
\arguments{
\item{db_map}{LIST object containing descriptions of table mapping in an
opinionated manner, generally generated by [er_map]. The expectation is a
list of tables, with references in SQL form enumerated in a child element
with a name matching `references_in`}

\item{references_in}{CHR scalar naming the child element containing SQL
references statements of the form "fk_column REFERENCES table(pk_column)"
(default: "references" is provided by [er_map])}

\item{dictionary}{LIST object containing the schema dictionary produced by
[data_dictionary] fully describing table entities}

\item{drop_if_exists}{LGL scalar indicating whether to include a "DROP VIEW"
prefix for the generated view statement; as this has an impact on schema,
no default is set}

\item{save_to_file}{CHR scalar name of a file path to save generated SQL
(default: NULL will return a list object to the R session)}

\item{append}{LGL scalar on whether to appead to `save_to_file` (default:
FALSE)}
}
\value{
LIST if `save_to_file = FALSE` or none
}
\description{
Build SQL to create views on normalized tables in SQLite
}
