% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementalcomposition.R
\name{calculate.monoisotope}
\alias{calculate.monoisotope}
\title{Calculate the monoisotopic mass of an elemental formula list}
\usage{
calculate.monoisotope(
  elementlist,
  exactmasses = NULL,
  adduct = "neutral",
  db_conn = "con"
)
}
\arguments{
\item{elementlist}{list of elemental formula from `extract.elements` function}

\item{exactmasses}{list of exact masses of elements}

\item{adduct}{character string adduct/charge state to add to the elemental formula, options are `neutral`, `+H`, `-H`, `+Na`, `+K`, `+`, `-`, `-radical`, `+radical`}

\item{db_conn}{database connection object, either a CHR scalar name (default: "con") or the connection object itself (preferred)}
}
\value{
numeric monoisotopic exact mass
}
\description{
Calculate the monoisotopic mass of an elemental formula list
}
\examples{
elementlist <- extract.elements("C2H5O")
calculate.monoisotope(elementalist, adduct = "neutral")

}
