% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{check_fragments}
\alias{check_fragments}
\title{Determine number of matching fragments between unknown mass spectrum and specific peaks}
\usage{
check_fragments(con, ums, peakid, masserror = 5, minerror = 0.001)
}
\arguments{
\item{con}{SQLite database connection}

\item{ums}{uncertainty mass spectrum of unknown compound}

\item{peakid}{integer vector of primary keys for peaks table}

\item{masserror}{numeric relative mass error (ppm)}

\item{minerror}{numeric minimum mass error (Da)}
}
\value{
table of fragments and TRUE/FALSE for if the fragment is within the unknown mass spectrum
}
\description{
Determine number of matching fragments between unknown mass spectrum and specific peaks
}
