% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopic_distribution.R
\name{check_isotopedist}
\alias{check_isotopedist}
\title{Compare Isotopic Pattern to simulated pattern}
\usage{
check_isotopedist(
  ms,
  elementalformula,
  exactmasschart,
  error,
  minerror = 0.002,
  remove.elements = c(),
  max.dist = 3,
  min.int = 0.001,
  charge = "neutral",
  m = 1,
  n = 0.5
)
}
\arguments{
\item{ms}{data.frame mass spectrum containing pair-wise m/z and intensity values of empirical isotopic pattern}

\item{elementalformula}{character string of elemental formula to simulate isotopic pattern}

\item{exactmasschart}{exact mass chart}

\item{error}{numeric relative mass error (in ppm) of mass spectrometer}

\item{minerror}{numeric minimum mass error (in Da) of mass spectrometer}

\item{remove.elements}{character vector of elements to remove from elemental formula}

\item{max.dist}{numeric maximum mass distance (in Da) from exact mass to include in simulated isotopic pattern}

\item{min.int}{numeric minimum relative intensity (maximum = 1, minimum = 0) to include in simulated isotopic pattern}

\item{charge}{character string for the charge state of the simulated isotopic pattern, options are `neutral`, `positive`, and `negative`}

\item{m}{numeric dot product mass weighting}

\item{n}{numeric dot product intensity weighting}
}
\value{
numeric vector of match scores between the empirical and calculated isotopic distribution.
}
\description{
calculates the isotopic distribution of the stated elemental formula and compares against the empirical ms
}
