% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{check_mzML_convert}
\alias{check_mzML_convert}
\title{Check mzML file for specific MSConvert parameters}
\usage{
check_mzML_convert(mzml)
}
\arguments{
\item{mzml}{list of msdata from `mzMLtoR` function}
}
\value{
data.frame object of conversion veracity checks
}
\description{
Check mzML file for specific MSConvert parameters
}
