% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_comm.R
\name{close_up_shop}
\alias{close_up_shop}
\title{Conveniently close all database connections}
\usage{
manage_connection()
close_up_shop(TRUE)
}
\arguments{
\item{back_up_connected_tbls}{LGL scalar of whether to clone currently
promised tibble connections to database objects as data frames (default:
FALSE).}
}
\value{
None, modifies the current global environment in place
}
\description{
This closes both the plumber service and all database connections from the
current running environment. If outstanding promises exist to database tables
or views were created as class `tbl_` (e.g. with `tbl(con, "table")`), set
`back_up_connected_tbls` to TRUE to collect data from those and preserve
in-place in the current global environment.
}
