% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_comparison.R
\name{compare_ms}
\alias{compare_ms}
\title{Calculate dot product match score}
\usage{
compare_ms(
  ms1,
  ms2,
  error = c(5, 5),
  minerror = c(0.002, 0.002),
  m = 1,
  n = 0.5
)
}
\arguments{
\item{ms1, ms2}{the uncertainty mass spectra from function `get_ums`}

\item{error}{a vector of the respective mass error (in ppm) for each mass spectrum or a single vector representing the mass error for all m/z values}

\item{minerror}{a two component vector of the respective minimum mass error (in Da) for each mass spectrum or a single value representing the minimum mass error of all m/z values}

\item{m, n}{weighting values for mass (m) and intensity (n)}
}
\description{
Calculates a the match score (based on dot product) of the two uncertainty mass spectra.
Note: this is a static match score and does not include associated uncertainties.
}
