% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions_shiny.R
\name{complete_form_entry}
\alias{complete_form_entry}
\title{Ensure complete form entry}
\usage{
req(complete_form_entry(input, c("need1", "need2")))
}
\arguments{
\item{input}{The session input object}

\item{values}{CHR vector of input object names to require}

\item{show_alert}{LGL scalar indicating whether or not to show an alert, set
FALSE to return the status of the check}
}
\value{
Whether or not all required values are present.
}
\description{
This input validation check ensures the current session's input object
includes non-NA, non-NULL, and non-blank values similarly to [shiny::req] and
[shiny::validate] but can be called with a predefined list of input names to
check. Typically this is used for validate form entry completion. Call this
function prior to reading form entries to ensure that all values requested by
name in in `values` are present. If they are not, a [nist_shinyalert] modal
is displayed prompting the user to complete the form.
}
