% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{create_peak_list}
\alias{create_peak_list}
\title{Create peak list from SQL ms_data table}
\usage{
create_peak_list(ms_data)
}
\arguments{
\item{ms_data}{extraction of the ms_data from the SQL table for a specified peak}
}
\value{
nested list of all data
}
\description{
The function extracts the relevant information and sorts it into nested lists for
use in the uncertainty functions
}
