% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{create_peak_table_ms2}
\alias{create_peak_table_ms2}
\title{Create peak table for MS2 data}
\usage{
create_peak_table_ms2(peak, mass, masserror = 5, minerror = 0.002, int0 = NA)
}
\arguments{
\item{mass}{the exact mass of the compound of interest}

\item{masserror}{the mass accuracy (in ppm) of the instrument data}

\item{minerror}{the minimum mass error (in Da) of the instrument data}

\item{int0}{the default setting for intensity values for missing m/z values}

\item{peaklist}{result of the `create_peak_list` function}
}
\value{
nested list of dataframes containing all MS2 data for the peak
}
\description{
Takes a nested peak list and creates a peak table for easier determination of 
uncertainty of the measurement for MS2 data.
}
