% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_setup.R
\name{create_py_env}
\alias{create_py_env}
\title{Create a python environment for RDKit}
\usage{
create_py_env("nist_hrms_db", c("reticulate", "rdkit"))
}
\arguments{
\item{env_name}{CHR scalar of a python environment}
}
\value{
None
}
\description{
This project offers a full integration of RDKit via [reticulate]. This
function does the heavy lifting for setting up that environment, either from
an environment specifications file or from the conda forge channel.
}
\details{
Preferred set up is to set variables in the `env_py.R` file, which will be
used over the internal defaults chosen here. The exception is if
`INSTALL_FROM == "local"` and no value is provided for `INSTALL_FROM_FILE`
which has no internal default.

Germane variables are `PYENV_NAME` (default "reticulated_rdkit"),
`CONDA_PATH` (default "auto"), `CONDA_MODULES` (default "rdkit",
"r-reticulate" will be added), `INSTALL_FROM` (default "conda"),
`INSTALL_FROM_FILE` (default "rdkit/environment.yml"), `MIN_PY_VER` (default
3.9).
}
