% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{create_search_df}
\alias{create_search_df}
\title{Create data.frame containing parameters for extraction and searching}
\usage{
create_search_df(
  filename,
  precursormz,
  rt,
  rt_start,
  rt_end,
  masserror,
  minerror,
  ms2exp,
  isowidth
)
}
\arguments{
\item{filename}{CHR scalar path to the mzml file}

\item{precursormz}{NUM scalar for the mass-to-charge ratio to examine}

\item{rt}{NUM scalar for the retention time centroid to examine}

\item{rt_start}{NUM scalar for the retention time start point of the feature}

\item{rt_end}{NUM scalar for the retention time end point of the feature}

\item{masserror}{NUM scalar of the instrument mass error value in parts per million}

\item{minerror}{NUM scalar of the minimum mass error value to use in absolute terms}

\item{ms2exp}{NUM scalar type of the fragmentation experiment (e.g. MS1 or MS2)}

\item{isowidth}{NUM scalar mass isolation width to use}
}
\value{
data.frame object collating provided values
}
\description{
Use this to create an intermediate data frame object used as part of the search routine.
}
