% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{create_search_ms}
\alias{create_search_ms}
\title{Generate uncertainty mass spectrum for MS1 and MS2 data}
\usage{
create_search_ms(
  searchobj,
  correl = NULL,
  ph = NULL,
  freq = NULL,
  normfn = "sum",
  cormethod = "pearson"
)
}
\arguments{
\item{searchobj}{list object generated from `get_search-object`}

\item{correl}{correlation limit for ions to MS1}

\item{ph}{peak height to select scans for generating mass spectrum}

\item{freq}{observational frequency minimum for ions to use for generating mass spectrum}

\item{normfn}{normalization function, options are "sum" or "mean"}

\item{cormethod}{correlation function, default is "pearson"}
}
\value{
list object containing the ms1 uncertainty mass spectrum `ums1`, ms2 uncertainty mass spectrum `ums2` and respective uncertainty mass spectrum parameters `ms1params` and `ms2params`
}
\description{
Generate uncertainty mass spectrum for MS1 and MS2 data
}
