% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_comm.R
\name{data_dictionary}
\alias{data_dictionary}
\title{Create a data dictionary}
\usage{
data_dictionary(db_conn = con)
}
\arguments{
\item{db_conn}{connection object (default:con)}
}
\value{
LIST of length equal to the number of tables in `con` with attributes
  identifying which tables, if any, failed to render into the dictionary.
}
\description{
Get a list of tables and their defined columns with properties, including
comments, suitable as a data dictionary from a connection object amenable to
[odbc::dbListTables]. This function relies on [pragma_table_info].
}
