% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_comparison.R
\name{dotprod}
\alias{dotprod}
\title{Calculate dot product}
\usage{
dotprod(m1, i1, m2, i2, m = 1, n = 0.5)
}
\arguments{
\item{m1, m2}{paired vectors containing measured m/z values}

\item{i1, i2}{paired vectors containing measured intensity values}

\item{m, n}{weighting values for mass (m) and intensity (n)}
}
\description{
Internal function: calculates the dot product between paired m/z and intensity values
}
