% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{dt_color_by}
\alias{dt_color_by}
\title{Apply colors to DT objects by value in a column}
\usage{
dt_color_by(names(DT_table_data), "color_by")
}
\arguments{
\item{table_names}{CHR vector of the names going into a table}

\item{look_for}{CHR vector of the column name to color by}
}
\value{
JS function to apply to a DT object by row
}
\description{
Adds a class to each node meeting the criteria defined elsewhere as project
object `table_bg_classes` as a list of colors with names matches values.
}
