% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_comm.R
\name{er_map}
\alias{er_map}
\title{Create a simple entity relationship map}
\usage{
er_map(db_conn = con)
}
\arguments{
\item{db_conn}{connection object, specifically of class "SQLiteConnection" but
not strictly enforced}
}
\value{
nested LIST object describing the database entity connections
}
\description{
This will poll the database connection and create an entity relationship map
as a list directly from defined SQL statements used to build the table or
view. For each table object it returns a list of length three containing the
entity names that the table (1) `references` (i.e. has a foreign key to), (2)
is `referenced_by` (i.e. is a foreign key for), and (3) views where it is
`used_in_view`. These are names. This is intended for use as a mapping
shortcut when ER Diagrams are unavailable, or for quick reference within a
project, similarly to a dictionary relationship reference.
}
\details{
SQL is generated from [pragma_table_def()] with argument `get_sql` = TRUE and
ignores entities whose names start with "sqlite".
}
