% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_msp.R
\name{export_msp}
\alias{export_msp}
\title{Export to MSP}
\usage{
export_msp(
  ms,
  file,
  precursor = "",
  name = "Exported Mass Spectrum",
  headerdata = c(),
  append = FALSE
)
}
\arguments{
\item{ms}{uncertainty mass spectrum from `get_ums` function}

\item{file}{export .msp file to save the msp files}

\item{precursor}{If available, the numeric precursor m/z for the designated mass spectrum}

\item{name}{Text name to assign to the mass spectrum (not used in spectral searching)}

\item{headerdata}{character string containing named values for additional data to put in the header}

\item{append}{boolean (TRUE/FALSE) to append to .msp file (TRUE) or overwrite (FALSE)}
}
\description{
The function exports an uncertainty mass spectrum into a NIST MS Search .msp file
}
