% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suspectlist.R
\name{extend_suspect_list}
\alias{extend_suspect_list}
\title{Extend the compounds and aliases tables}
\usage{
extend_suspect_list(suspect_list, db_conn = con, retain_current = TRUE)
}
\arguments{
\item{suspect_list}{CHR scalar pointing either to a file (CSV, XLS, or XLSX)
or URL pointing to an XLSX file.}

\item{db_conn}{connection object (default: con)}

\item{retain_current}{LGL scalar of whether to retain the current list by
attempting to match new entries to older ones, or to append all entries
(default: TRUE)}
}
\value{
None
}
\description{
Suspect lists are occasionally updated. To keep the current database up to
date, run this function by pointing it to the updated or current suspect
list. That suspect list should be one of (1) a file in either
comma-separated-value (CSV) or a Microsoft Excel format (XLS or XLSX), (2) a
data frame containing the new compounds in the standard format of the suspect
list, or (3) a URL pointing to the suspect list.
}
\details{
If `suspect_list` does not contain one of the expected file extensions, it
will be assumed to be a URL pointing to a Microsoft Excel file with the
suspect list in the first spreadsheet. The file for that URL will be
downloaded temporarily, read in as a data frame, and then removed.

Required columns for the compounds table are first pulled and all other
columns are treated as aliases. If `retain_current` is TRUE, entries in the
"name" column will be matched against current aliases and the compound id
will be persisted for that compound.
}
