% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementalcomposition.R
\name{extract.elements}
\alias{extract.elements}
\title{Elemental Formula Functions
Extract elements from formula}
\usage{
extract.elements(composition.str, remove.elements = c())
}
\arguments{
\item{composition.str}{character string elemental formula}

\item{remove.elements}{character vector containing elements to remove from}
}
\value{
list with `elements` and `counts`
}
\description{
Converts elemental formula into list of `elements` and `counts` corresponding to the composition
}
\examples{
extract.elements("C2H5O")

extract.elements("C2H5ONa", remove.elements = c("Na", "Cl"))
}
