% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{flush_dir}
\alias{flush_dir}
\title{Flush a directory with archive}
\usage{
flush_dir("logs", ".txt")

flush_dir(directory = "logs")
}
\arguments{
\item{archive}{LGL scalar on whether to archive current logs}

\item{directory}{CHR scalar path to the directory to flush}
}
\value{
None, executes directory actions

None, removes files from a directory
}
\description{
Clear a directory and archive those files if desired in any directory
matching any pattern.

Clear a directory and archive those files if desired in any directory
matching any pattern.
}
