% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{fn_help}
\alias{fn_help}
\title{Get function documentation for this project}
\usage{
fn_help(fn_name)
}
\arguments{
\item{fn_name}{Object or CHR string name of a function in this project.}
}
\value{
None, opens help file.
}
\description{
This function is analogous to "?", "??", and "help". For now, this effort is
distributed as a project instead of a package. This imposes certain
limitations, particularly regarding function documentation. Use this function
to see the documentation for functions in this project just as you would any
installed package. The other limitation is that these help files will not
populate directly as a pop up when using RStudio tab completion.
}
\note{
This function will be deprecated if the project is moved to a package.
}
\examples{
fn_help(fn_help)
}
