% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{format_html_id}
\alias{format_html_id}
\title{Format a file name as an HTML element ID}
\usage{
format_html_id(filename)
}
\arguments{
\item{filename}{CHR vector of file names}
}
\value{
CHR vector of the same size as filename
}
\description{
This is often useful to provide feedback to the user about the files they've
provided to a shiny application in a more informative manner, as IDs produced
here are suitable to build dynamic UI around. This can serve as the base ID
for tooltips, additional information, icons, etc. and produce everything
necessary in one place for any number of files.
}
\examples{
format_html_id(list.files())

}
