% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{format_list_of_names}
\alias{format_list_of_names}
\title{Grammatically collapse a list of values}
\usage{
format_list_of_names(namelist, add_quotes = FALSE)
}
\arguments{
\item{namelist}{vector of values to format}

\item{add_quotes}{LGL scalar of whether to enclose individual values in
quotation marks}
}
\value{
CHR vector of length one
}
\description{
Given a vector of arbitrary length that coerces properly to a human-readable
character string, return it formatted as one of: "one", "one and two", or
"one, two, ..., and three" using \code{glue::glue}. This is functionally the same
as a static version of [glue::glue_collapse] with parameters sep = ", ",
width = Inf, and last = ", and ".
}
\examples{
format_list_of_names("test")
format_list_of_names(c("apples", "bananas"))
format_list_of_names(c(1:3))
format_list_of_names(seq.Date(Sys.Date(), Sys.Date() + 3, by = 1))
}
