% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulalize.R
\name{formulalize}
\alias{formulalize}
\title{Generate standard chemical formula notation}
\usage{
formulalize(formula)
}
\arguments{
\item{formula}{CHR string of an elemental formula}
}
\value{
string with a standard ordered formula
}
\description{
Generate standard chemical formula notation
}
\examples{

formula <- "C10H15S1O3"
formulalize(formula)
}
