% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_qc.R
\name{gather_qc}
\alias{gather_qc}
\title{Quality Control Check of Import Data}
\usage{
gather_qc(
  gather_peak,
  exactmasses,
  ms1range = c(0.5, 3),
  ms1isomatchlimit = 0.5,
  minerror = 0.002
)
}
\arguments{
\item{gather_peak}{peak object generated from `peak_gather_json` function}

\item{exactmasses}{exactmasses list}

\item{ms1range}{2-component vector containing stating the range to evaluate the isotopic pattern of the precursor ion, from mass - ms1range[1] to mass + ms1range[2]}

\item{ms1isomatchlimit}{the reverse dot product minimum score for the isotopic pattern match}

\item{minerror}{the minimum mass error (in Da) allowable for the instrument}
}
\value{
nested list of quality control check results
}
\description{
Performs the quality control check on the imported data from the peak gather function.
}
