% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_annotated_fragments}
\alias{get_annotated_fragments}
\title{Get all annotated fragments have matching masses}
\usage{
get_annotated_fragments(con, fragmentions, masserror, minerror)
}
\arguments{
\item{con}{SQLite database connection}

\item{fragmentions}{numeric vector containing m/z values for fragments to search}

\item{masserror}{numeric relative mass error (ppm)}

\item{minerror}{numeric minimum mass error (Da)}
}
\value{
data.frame of mass spectral data
}
\description{
Get all annotated fragments have matching masses
}
