% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_compound_fragments}
\alias{get_compound_fragments}
\title{Get all fragments associated with compounds}
\usage{
get_compound_fragments(con, fragmentions, masserror, minerror)
}
\arguments{
\item{con}{SQLite database connection}

\item{fragmentions}{numeric vector containing m/z values for fragments to search}

\item{masserror}{numeric relative mass error (ppm)}

\item{minerror}{numeric minimum mass error (Da)}
}
\value{
data.frame object describing known fragments in the database with known compound and peak references attached
}
\description{
Get all fragments associated with compounds
}
