% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_compoundid}
\alias{get_compoundid}
\title{Get compound ID and name for specific peaks}
\usage{
get_compoundid(con, peakid)
}
\arguments{
\item{con}{SQLite database connection}

\item{peakid}{integer vector of primary keys for peaks table}
}
\value{
table of compound IDs and names
}
\description{
Get compound ID and name for specific peaks
}
