% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_msdata}
\alias{get_msdata}
\title{Get all mass spectral data within the database}
\usage{
get_msdata(con)
}
\arguments{
\item{con}{SQLite database connection}
}
\value{
data.frame of mass spectral data
}
\description{
Get all mass spectral data within the database
}
