% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_msdata_compound}
\alias{get_msdata_compound}
\title{Get all mass spectral data for a specific compound}
\usage{
get_msdata_compound(con, 15)
}
\arguments{
\item{con}{SQLite database connection}

\item{compoundid}{integer compound ID value}
}
\value{
data.frame of mass spectral data
}
\description{
Get all mass spectral data for a specific compound
}
