% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_msdata_peakid}
\alias{get_msdata_peakid}
\title{Get all mass spectral data for a specific peak id}
\usage{
get_msdata_peakid(con, 15)
}
\arguments{
\item{con}{SQLite database connection}

\item{peakid}{integer vector of peak ids}
}
\value{
data.frame of mass spectral data
}
\description{
Get all mass spectral data for a specific peak id
}
