% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_msdata_precursors}
\alias{get_msdata_precursors}
\title{Get all mass spectral data with a specific precursor ion}
\usage{
get_msdata_precursors(con, precursorion, masserror, minerror)
}
\arguments{
\item{con}{SQLite database connection}

\item{precursorion}{numeric precursor ion m/z value}

\item{masserror}{numeric relative mass error (ppm)}

\item{minerror}{numeric minimum mass error (Da)}
}
\value{
data.frame of mass spectral data
}
\description{
Get all mass spectral data with a specific precursor ion
}
