% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_ums.R
\name{get_opt_params}
\alias{get_opt_params}
\title{Get optimized uncertainty mass spectra parameters for a peak}
\usage{
get_opt_params(con, peak_ids)
}
\arguments{
\item{con}{SQLite database connection}

\item{peak_ids}{integer vector of primary keys for peaks table}
}
\value{
data.frame object of available optimized search parameters
}
\description{
Get optimized uncertainty mass spectra parameters for a peak
}
