% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_peak_fragments}
\alias{get_peak_fragments}
\title{Get annotated fragments for a specific peak}
\usage{
get_peak_fragments(con, peakid)
}
\arguments{
\item{con}{SQLite database connection}

\item{peakid}{integer vector of primary keys for peaks table}
}
\value{
data.frame of annotated fragments
}
\description{
Get annotated fragments for a specific peak
}
