% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_peak_precursor}
\alias{get_peak_precursor}
\title{Get precursor ion m/z for a specific peak}
\usage{
get_peak_precursor(con, peakid)
}
\arguments{
\item{con}{SQLite database connection}

\item{peakid}{integer primary key for peaks table}
}
\value{
numeric value of precursor ion m/z value
}
\description{
Get precursor ion m/z for a specific peak
}
