% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_sample_class}
\alias{get_sample_class}
\title{Get sample class information for specific peaks}
\usage{
get_sample_class(con, peakid)
}
\arguments{
\item{con}{SQLite database connection}

\item{peakid}{integer vector of primary keys for peaks table}
}
\value{
data.frame object of sample classes associated with a given peak
}
\description{
Get sample class information for specific peaks
}
