% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suspectlist.R
\name{get_suspectlist}
\alias{get_suspectlist}
\title{Get the current NIST PFAS suspect list.}
\usage{
get_suspectlist(
  destfile = file.path("R", "compoundlist", "suspectlist.xlsx"),
  url_file = file.path("config", "suspectlist_url.txt"),
  default_url = SUS_LIST_URL,
  save_local = FALSE
)
}
\arguments{
\item{destfile}{CHR scalar file.path of location to save the downloaded file}

\item{url_file}{CHR scalar file.path of where the text file containing the
download URL for the NIST PFAS Suspect List}

\item{save_local}{LGL scalar of whether to retain an R expression in the
current environment after download}
}
\value{
none
}
\description{
Downloads the current NIST suspect list of PFAS from the NIST Public Data
Repository to the current project directory.
}
\examples{
get_suspectlist()
}
