% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{get_uniques}
\alias{get_uniques}
\title{Get unique components of a nested list}
\usage{
get_uniques(objects, aspect)
}
\arguments{
\item{objects}{LIST object}

\item{aspect}{CHR scalar name of the aspect from which to generate unique
combinations}
}
\value{
Unnamed LIST of length equaling the number of unique combinations
  with their values and indices
}
\description{
There are times when the concept of "samples" and "grouped data" may become
intertwined and difficult to parse. The import process is one of those times
depending on how the import file is generated. This function takes a nested
list and compares a specific aspect of it, grouping the output based on that
aspect and returning its characteristics.
}
\details{
For example, the standard NIST import includes the "sample" aspect, which may
be identical for multiple data import files. This provides a unique listing
of those sample characteristics to reduce data manipulation and storage, and
minimize database "chatter" during read/write. It returns a set of unique
characteristics in a list, with appended characteristics "import_object" with
the index number and object name of entries matching those characteristics.

This is largely superceded by later developments to database operations that
first check for a table primary key id given a comprehensive list of column
values in those tables where only a single record should contain those values
(e.g. a complete unique case, enforced or unenforced).
}
\examples{
tmp <- list(list(a = 1:10, b = 1:10), list(a = 1:5, b = 1:5), list(a = 1:10, b = 1:5))
get_uniques(tmp)
}
