% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML_base.R
\name{getcharge}
\alias{getcharge}
\title{Get polarity of a ms scan within mzML object}
\usage{
getcharge(mzml, i)
}
\arguments{
\item{mzml}{list mzML object generated from `mzMLtoR` function}

\item{i}{integer scan number}
}
\value{
integer representing scan polarity (either 1 (positive) or -1 (negative))
}
\description{
Get polarity of a ms scan within mzML object
}
