% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML_base.R
\name{getmslevel}
\alias{getmslevel}
\title{Get MS Level of a ms scan within mzML object}
\usage{
getmslevel(mzml, i)
}
\arguments{
\item{mzml}{list mzML object generated from `mzMLtoR` function}

\item{i}{integer scan number}
}
\value{
integer representing the MS Level (1, 2, ... n)
}
\description{
Get MS Level of a ms scan within mzML object
}
