% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{getmzML}
\alias{getmzML}
\title{Brings raw data file into environment}
\usage{
getmzML(
  search_df,
  CONVERT = FALSE,
  CHECKCONVERT = TRUE,
  is_waters = FALSE,
  lockmass = NULL,
  lockmasswidth = NULL,
  correct = FALSE
)
}
\arguments{
\item{search_df}{data.frame output of [create_search_df] or file name of a raw file to be converted}

\item{CONVERT}{LGL scalar of whether or not to convert the search_df filename (default FALSE)}

\item{CHECKCONVERT}{LGL scalar of whether or not to verify the conversion format (default TRUE)}
}
\value{
LIST value of the trimmed mzML file matching search criteria
}
\description{
If filename is not extension .mzML, then converts the raw file
}
