% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML_base.R
\name{getprecursor}
\alias{getprecursor}
\title{Get precursor ion of a ms scan within mzML object}
\usage{
getprecursor(mzml, i)
}
\arguments{
\item{mzml}{list mzML object generated from `mzMLtoR` function}

\item{i}{integer scan number}
}
\value{
numeric designating the precursor ion (or middle of the scan range for SWATCH or DIA), returns NULL if no precursor was selected
}
\description{
Get precursor ion of a ms scan within mzML object
}
