% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML_base.R
\name{gettime}
\alias{gettime}
\title{Get time of a ms scan within mzML object}
\usage{
gettime(mzml, i)
}
\arguments{
\item{mzml}{list mzML object generated from `mzMLtoR` function}

\item{i}{integer scan number}
}
\value{
numeric of the scan time
}
\description{
Get time of a ms scan within mzML object
}
