% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{has_missing_elements}
\alias{has_missing_elements}
\title{Simple check for if an object is empty}
\usage{
has_missing_elements(x, logging = TRUE)
}
\arguments{
\item{x}{Object to be checked}

\item{logging}{LGL scalar of whether or not to make log messages (default: TRUE)}
}
\value{
LGL scalar of whether \code{x} is empty
}
\description{
Checks for empty vectors, a blank character string, NULL, and NA values. If
fed a list object, returns TRUE if any element is is the "empty" set. For
data.frames checks that nrow is not 0. [rlang:::is_empty] only checks for
length 0.
}
\note{
Reminder that vectors created with NULL values will be automatically
  reduced by R.
}
\examples{
has_missing_elements("a")
# FALSE
has_missing_elements(c(NULL, 1:5))
# FALSE
has_missing_elements(list(NULL, 1:5))
# TRUE
has_missing_elements(data.frame(a = character(0)))
# TRUE
}
