% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementalcomposition.R
\name{is_elemental_match}
\alias{is_elemental_match}
\title{Checks if two elemental formulas match}
\usage{
is_elemental_match(testformula, trueformula)
}
\arguments{
\item{testformula}{character string of elemental formula to test}

\item{trueformula}{character string of elemental formula to check against (truth)}
}
\value{
logical
}
\description{
Checks if two elemental formulas match
}
