% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementalcomposition.R
\name{is_elemental_subset}
\alias{is_elemental_subset}
\title{Check if elemental formula is a subset of another formula}
\usage{
is_elemental_subset(fragmentformula, parentformula)
}
\arguments{
\item{fragmentformula}{character string of elemental formula subset to test}

\item{parentformula}{character string of elemental formula to check for subset}
}
\value{
logical
}
\description{
Check if elemental formula is a subset of another formula
}
\examples{
is_elemental_subset("C2H2", "C2H5O")

is_elemental_subset("C2H2", "C2H1O")
}
