% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopic_distribution.R
\name{isotopic_distribution}
\alias{isotopic_distribution}
\title{Isotopic distribution functions
Generate isotopic distribution mass spectrum of elemental formula}
\usage{
isotopic_distribution(
  elementalformula,
  exactmasschart,
  remove.elements = c(),
  max.dist = 3,
  min.int = 0.001,
  charge = "neutral"
)
}
\arguments{
\item{elementalformula}{}

\item{exactmasschart}{}

\item{remove.elements}{}

\item{max.dist}{}

\item{min.int}{}

\item{charge}{}
}
\value{

}
\description{
Isotopic distribution functions
Generate isotopic distribution mass spectrum of elemental formula
}
