% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML_base.R
\name{lockmass_remove}
\alias{lockmass_remove}
\title{Remove lockmass scan from mzml object}
\usage{
lockmass_remove(
  mzml,
  lockmass = NULL,
  lockmasswidth = NULL,
  correct = FALSE,
  approach = "baseion"
)
}
\arguments{
\item{mzml}{mzML object generated from mzMLtoR() function}

\item{lockmass}{m/z value of the lockmass to remove}

\item{lockmasswidth}{m/z value for the half-window of the lockmass scan}

\item{correct}{logical if the subsequent spectra should be corrected}
}
\value{
A copy of the object provided to `mzml` with the lock mass removed.
}
\description{
For Waters instruments only, identifies the scans that are due to a lock mass scan
and removes them for easier processing.
}
