% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{log_fn}
\alias{log_fn}
\title{Simple logging convenience}
\usage{
fn <- function() {log_fn("start"); 1+1; log_fn("end")}
fn()
}
\arguments{
\item{status}{CHR scalar to prefix the log message; will be coerced to
sentence case. Typically "start" or "end" but anything is accepted (default
"start").}

\item{log_ns}{CHR scalar of the logger namespace to use (default NA_character_)}

\item{level}{CHR scalar of the logging level to be passed to [log_it]
(default "trace")}
}
\value{
None, hands logging messages to [log_it]
}
\description{
Conveniently add a log message at the trace level. Typically this would be
called twice bookending the body of a function along the lines of "Start
fn()" and "End fn()" when calling a function. This can help provided
traceability to deeply nested function calls within a log.
}
