% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{make_acronym}
\alias{make_acronym}
\title{Simple acronym generator}
\usage{
make_acronym(text)
}
\arguments{
\item{text}{CHR vector of the text to acronym-ize}
}
\value{
CHR vector of length equal to that of \code{text} with the acronym
}
\description{
At times it is useful for display purposes to generate acronyms for longer
bits of text. This naively generates those by extracting the first letter as
upper case from each word in \code{text} elements.
}
\examples{
make_acronym("test me")
make_acronym(paste("department of ", c("commerce", "energy", "defense")))
}
