% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_comm.R
\name{make_install_code}
\alias{make_install_code}
\title{Convenience function to set a new installation code}
\usage{
make_install_code(db_conn = con, new_name = NULL, log_ns = "db")
}
\arguments{
\item{db_conn}{connection object (default "con")}

\item{new_name}{CHR scalar of the human readable name of the installation
(e.g. your project name) (default: NULL)}

\item{log_ns}{CHR scalar of the logging namespace to use during execution
(default: "db")}
}
\value{
None
}
\description{
Convenience function to set a new installation code
}
