% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{make_requirements}
\alias{make_requirements}
\title{Make import requirements file}
\usage{
make_requirements(
  example_import,
  file_name = "import_requirements.json",
  not_required = c("annotation", "chromatography", "opt_ums_params"),
  archive = TRUE,
  retain_in_R = TRUE,
  log_ns = "db"
)
}
\arguments{
\item{example_import}{CHR or LIST object containing an example of the
expected import format; this should include only a SINGLE compound
contribution file}

\item{file_name}{CHR scalar indicating a file name to save the resulting name
or search on any existing file to archive if `archive` = TRUE (default:
"import_requirements.json")}

\item{not_required}{CHR vector matching element names of `example_import`
which are not required; all others will be assumed to be required}

\item{archive}{LGL indicating whether or not to archive an existing file
matching `file_name` by suffixing the file name with current date. Only one
archive per date is supported; if a file already exists, it will be
deleted. (default: TRUE)}

\item{retain_in_R}{LGL indicating whether to retain a local copy of the
requirements file generated (default: TRUE)}

\item{log_ns}{CHR scalar of the logging namespace to use (default: "db")}
}
\value{
writes a file to the project directory (based on the found location
  of `file_name`) with the JSON structure
}
\description{
Importing from the NIST contribution spreadsheet requires a certain format.
In order to proceed smoothly, that format must be verified for gross
integrity with regard to expectations about shape (i.e. class), names of
elements, and whether they are required for import. This function creates a
JSON expression of the expected import structure and saves it to the project
directory.
}
\details{
Either an existing JSON expression or an R list object may be used for
`example_import`. If it is a character scalar, it will be assumed to be a
file name, which will be loaded based on file extension. That file must be a
JSON parseable text file, though raw text is acceptable.

An example file is located in the project directory at
"example/PFAC30PAR_PFCA1_mzML_cmpd2627.JSON"

As with any file manipulation, use care with `file_name`.
}
