% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_setup.R
\name{molecule_picture}
\alias{molecule_picture}
\title{Picture a molecule from structural notation}
\usage{
caffeine <- "C[n]1cnc2N(C)C(=O)N(C)C(=O)c12"
molecule_picture(caffeine, show = TRUE)
}
\arguments{
\item{mol}{CHR scalar expression of molecular structure}

\item{mol_type}{CHR scalar indicating the expression type of `mol` (default:
"smiles")}

\item{file_name}{CHR scalar of an intended file destination (default: NULL
will produce a random 10 character file name). Note that any file
extensions provided here will be ignored.}

\item{rdkit_name}{CHR scalar indication the name of the R object bound to
RDkit OR the name of the R object directly (i.e. without quotes)}

\item{open_file}{LGL scalar of whether to open the file after creation
(default: FALSE)}

\item{show}{LGL scalar of whether to return the image itself as an object
(default: FALSE)}
}
\value{
None, or displays the resulting picture if `show == TRUE`
}
\description{
This is a thin wrapper to rdkit.Chem.MolFromX methods to generate molecular
models from common structure notation such as InChI or SMILES. All picture
files produced will be in portable network graphics (.png) format.
}
\note{
Supported `mol` expressions include FASTA, HELM, Inchi, Mol2Block,
  Mol2File, MolBlock, MolFile, PDBBlock, PDBFile, PNGFile, PNGString,
  RDKitSVG, Sequence, Smarts, Smiles, TPLBlock, and TPLFile
}
