% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementalcomposition.R
\name{monoisotope.list}
\alias{monoisotope.list}
\title{Calculate the monoisotopic mass of a elemental formulas in}
\usage{
monoisotope.list(
  df,
  column,
  exactmasses,
  remove.elements = c(),
  adduct = "neutral"
)
}
\arguments{
\item{df}{data.frame with at least one column with elemental formulas}

\item{column}{integer or CHR scalar indicating the column containing the elemental formulas, if CHR then regex match is used}

\item{exactmasses}{list of exact masses}

\item{remove.elements}{elements to remove from the elemental formulas}

\item{adduct}{character string adduct/charge state to add to the elemental formula, options are `neutral`, `+H`, `-H`, `+Na`, `+K`, `+`, `-`, `-radical`, `+radical`}
}
\value{
data.frame with column of exact masses appended to it
}
\description{
Calculate the monoisotopic mass of a elemental formulas in
}
