% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_spectra.R
\name{ms_plot_peak}
\alias{ms_plot_peak}
\title{Plot a peak from database mass spectral data}
\usage{
ms_plot_peak(
  data,
  peak_type = c("area", "line", "segment"),
  peak_facet_by = "ms_n",
  peak_mz_resolution = 0,
  peak_drop_ratio = 0.01,
  peak_repel_labels = TRUE,
  peak_line_color = "black",
  peak_fill_color = "grey50",
  peak_fill_alpha = 0.2,
  peak_text_size = 3,
  peak_text_offset = 0.02,
  include_method = TRUE,
  db_conn = con
)
}
\arguments{
\item{data}{data.frame of spectral data in the form of the `ms_data` table}

\item{peak_type}{CHR scalar of the plot type to draw, must be one of "line",
"segment", or "area" (default: "line")}

\item{peak_facet_by}{CHR scalar name of a column by which to facet the
resulting plot (default: "ms_n")}

\item{peak_mz_resolution}{INT scalar mass to charge ratio tolerance to group
peaks, with at minimum columns for intensity (as "base_int"), ion m/z value
(as "base_ion"), and scan time (as "scantime") - (default: 0 goes to unit
resolution)}

\item{peak_drop_ratio}{NUM scalar threshold of the maximum intensity below
which traces will be dropped (default: 1e-2 means any trace with a maximum
intensity less than 1\% of the maximum intensity in the plot will be
dropped); if > 1 the inversion will be used (1e5 -> 1e-5)}

\item{peak_repel_labels}{LGL scalar on whether to use the [ggrepel] package
to space out m/z labels in the plot (default: TRUE). If [ggrepel] is not
installed, it will default to FALSE rather than requiring an installation}

\item{peak_line_color}{CHR scalar name of the color to use for the "color"
aesthetic (only a single color is supported; default: "black")}

\item{peak_fill_color}{CHR scalar name of the color to use for the "fill"
aesthetic (only a single color is supported; default: "grey70")}

\item{peak_text_offset}{NUM scalar y-axis offset as a fraction of the maximum
intensity for trace annotation (default: 0.02 offsets labels in the
positive direction by 2\% of the maximum intensity)}

\item{db_conn}{database connection (default: con) which must be live to pull
sample and compound identification information}
}
\value{
ggplot object
}
\description{
Plots the intensity of ion traces over the scan period and annotates them
with the mass to charge value. Several flexible plotting aspects are provided
as data may become complicated.
}
\details{
The basic default plot will group all mass-to-charge ratio values by unit
resolution (increase resolution with `peak_mz_resolution`) and plot them as
an area trace over the scanning period. Traces are annotated with the
grouping value. Values of `peak_mz_resolution` greater than available data
(e.g. 10 when data resolution is to the 5th decimal point) will default to
maximum resolution.

Traces are filtered out completely if their maximum intensity is below the
ratio set by `peak_drop_ratio`; only complete traces are filtered out this
way, not individual data points within a retained trace. Set this as the
fraction of the base peak (the peak of maximum intensity) to use to filter
out low-intensity traces. The calculated intensity threshold will be printed
to the caption.
}
\note{
Increasing `peak_mz_resolution` will likely result in multiple separate
  traces.

Implicitly missing values are not interpolated, but lines are drawn
  through to the next point.

`peak_type` can will accept abbreviations of its accepted values (e.g.
  "l" for "line")
}
