% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_spectra.R
\name{ms_plot_spectral_intensity}
\alias{ms_plot_spectral_intensity}
\title{Create a spectral intensity plot}
\usage{
ms_plot_spectral_intensity(
  data,
  intensity_mz_resolution = 5,
  intensity_drop_ratio = 0,
  intensity_facet_by = NULL,
  intensity_plot_resolution = c("spectra", "peak"),
  include_method = TRUE,
  db_conn = con
)
}
\arguments{
\item{data}{tibble or pointer with data to plot, either at the peak level, in
which case "base_ion" must be present, or at the spectral level, in which
case "intensity" must be present}

\item{intensity_mz_resolution}{INT scalar mass to charge ratio tolerance to
group peaks, with at minimum columns for intensity (as "base_int" or
"intensity"), ion m/z value (as "base_ion" or "mz"), and scan time (as
"scantime") - (default: 5)}

\item{intensity_drop_ratio}{NUM scalar threshold of the maximum intensity
below which traces will be dropped (default: 0 returns all); if > 1 the
inversion will be used (1e5 -> 1e-5)}

\item{intensity_facet_by}{CHR scalar of a column name in `data` by which to
facet the resulting plot (default: NULL)}

\item{db_conn}{database connection (default: con) which must be live to pull
sample and compound identification information}
}
\value{
object of classes `gg` and `ggplot`
}
\description{
Often it is useful to get an overview of mass-to-charge intensity across the
scanning time of a peak. Typically this is done with individual traces in the
peak fashion, but large peaks can often mask smaller ones, or wash out lower
intensity signals. Use this to plot m/z as dependent upon scan time with
intensity shown by color and size. It is intended as a complement to
[ms_plot_peak] and may be called at the same levels of granularity, generally
greater so than [ms_plot_peak] which is more of an overview.
}
