% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_spectra.R
\name{ms_plot_titles}
\alias{ms_plot_titles}
\title{Consistent title for ms_plot_x functions}
\usage{
ms_plot_titles(
  plot_data,
  mz_resolution,
  drop_ratio,
  include_method,
  db_conn = con
)
}
\arguments{
\item{plot_data}{data.frame object passed from the plotting function}

\item{mz_resolution}{NUM scalar passed from the plotting function}

\item{drop_ratio}{NUM scalar passed from the plotting function}

\item{include_method}{LGL scalar indicating whether or not to get the method
narrative from the database}

\item{db_conn}{database connection (default: con) which must be live to pull
sample and compound identification information}
}
\value{
LIST of strings named for ggplot title elements "title", "subtitle",
  and "caption"
}
\description{
This helper function creates consistently formatted plot label elements in an
opinionated manner. This is unlikely to be useful outside the direct context
of [ms_plot_peak], [ms_plot_spectra], and [ms_plot_spectral_intensity].
}
