% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_spectra.R
\name{ms_spectra_separated}
\alias{ms_spectra_separated}
\title{Parse "Separated" MS Data}
\usage{
ms_spectra_separated(df, ms_cols = c("mz", "intensity"))
}
\arguments{
\item{df}{data.frame or json object containing spectra compressed in the
"separated" format}

\item{ms_cols}{CHR vector of length 2 identifying the column names to use for
mass and intensity in the source data; must be of length 2, with the first
value identifying the mass-to-charge ratio column and the second
identifying the intensity column}
}
\value{
data.frame object of the unpacked spectra as a list column
}
\description{
The "separated" format includes spectra packed into two separate columns, one
for mass and another for intensity. All values for a given scan time are
packed into these columns, separated by space, with an unlimited number of
discrete values, and must be a 1:1 ratio of values between the two columns.
}
\note{
ms_cols is treated as regex expressions, but it is safest to provide
  matching column names
}
\examples{
### JSON Example
tmp <- jsonify::as.json('{
 "measured_mz": "712.9501 713.1851",
 "measured_intensity": "15094.41015625 34809.9765625"
}')
ms_spectra_separated(tmp)

### Example data.frame
tmp <- data.frame(
  measured_mz = "712.9501 713.1851",
  measured_intensity = "15094.41015625 34809.9765625"
)
ms_spectra_separated(tmp)
}
