% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_spectra.R
\name{ms_spectra_zipped}
\alias{ms_spectra_zipped}
\title{Parse "Zipped" MS Data}
\usage{
ms_spectra_zipped(df, spectra_col = "data")
}
\arguments{
\item{df}{data.frame object containing spectra compressed in the "zipped"
format}

\item{spectra_col}{CHR vector of length 2 identifying the column names to use
for mass and intensity in the source data; must be of length 2, with the
first value identifying the mass column and the second identifying the
intensity column}
}
\value{
data.frame object containing unpacked spectra as a list column
}
\description{
The "zipped" format includes spectra packed into one column containing
alternating mass and intensity values for all observations. All values are
packed into these columns for a given scan time, separated by spaces, with an
unlimited number of discrete values, and must be in an alternating 1:1
pattern of values of the form "mass intensity mass intensity".
}
\note{
spectra-col is treated as a regex expression, but it is safest to
  provide a matching column name
}
\examples{
### JSON Example
tmp <- jsonlite::as.json('{"msdata": "712.9501 15094.41015625 713.1851 34809.9765625"}')
ms_spectra_separated(tmp)

### Example data.frame
tmp <- data.frame(
  msdata = "712.9501 15094.41015625 713.1851 34809.9765625"
)
ms_spectra_zipped(tmp)
}
