% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzmlconvert.R
\name{mzMLconvert}
\alias{mzMLconvert}
\title{Converts a raw file into an mzML}
\usage{
mzMLconvert(rawfile, msconvert = NULL, config = NULL, outdir = getwd())
}
\arguments{
\item{rawfile}{file path of the MS raw file to be converted}

\item{msconvert}{file path of the msconvert.exe file, if NULL retrieves information from config directory}

\item{config}{configuration settings file for msconvert conversion to mzML, if NULL retrives information from config directory}

\item{outdir}{directory path for the converted mzML file.}
}
\value{
CHR scalar path to the created file
}
\description{
Converts a raw file into an mzML
}
