% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML_base.R
\name{mzMLtoR}
\alias{mzMLtoR}
\title{Opens file of type mzML into R environment}
\usage{
mzMLtoR(mzmlfile = file.choose())
}
\arguments{
\item{mzmlfile}{the name of the mzML file which the data are to be read from.}
}
\value{
list containing mzML data with unzipped masses and intensity information
}
\description{
Opens file of type mzML into R environment
}
\examples{

}
