% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions_shiny.R
\name{nist_shinyalert}
\alias{nist_shinyalert}
\title{Call [shinyalert::shinyalert] with specific styling}
\usage{
nist_shinyalert("test", "info", shiny::h3("test"))
}
\arguments{
\item{title}{The title of the modal.}

\item{type}{The type of the modal. There are 4 built-in types which will show
a corresponding icon: \code{"warning"}, \code{"error"}, \code{"success"} and
\code{"info"}. You can also set \code{type="input"} to get a prompt
in the modal where the user can enter a response. By default, the modal has
no type.}

\item{text}{The modal's text. Can either be simple text, or Shiny tags (including
Shiny inputs and outputs). If using Shiny tags, then you must also set \code{html=TRUE}.}

\item{className}{A custom CSS class name for the modal's container.}

\item{html}{If \code{TRUE}, the content of the title and text will not be
escaped. By default, the content in the title and text are escaped, so any
HTML tags will not render as HTML.}

\item{closeOnClickOutside}{If \code{TRUE}, the user can dismiss the modal by
clicking outside it.}

\item{immediate}{If \code{TRUE}, close any previously opened alerts and display
the current one immediately.}

\item{...}{Additional named parameters to be passed to shinyalert.
Unrecognized ones will be ignored.}
}
\value{
None, shows a shinyalert modal
}
\description{
This pass through function serves only to call [shinyalert::shinyalert] with
parameters defined by this function, and can be used for additional styling
that may be necessary. It is used solely for consistency sake.
}
\seealso{
shinyalert::shinyalert
}
