% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions_devonly.R
\name{open_env}
\alias{open_env}
\title{Convenience shortcut to open and edit session environment variables}
\usage{
open_env(name = c("R", "global", "logging", "rdkit", "shiny", "plumber"))
}
\arguments{
\item{name}{CHR scalar, one of "R", "global", or "logging".}
}
\value{
None, opens a file for editing
}
\description{
Calls [open_proj_file] for either the R, global, or logging environment
settings containing the most common settings dictating project behavior.
}
