% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions_devonly.R
\name{open_proj_file}
\alias{open_proj_file}
\title{Open and edit project files}
\usage{
open_proj_file(name, dir = NULL, create_new = FALSE)
}
\arguments{
\item{name}{CHR scalar of the file name to open, accepts regex}

\item{dir}{CHR scalar of a directory name to search within}

\item{create_new}{LGL scalar of whether to create the file (similar
functionality to [usethis]; default FALSE)}
}
\value{
None, opens a file for editing
}
\description{
Project files are organized in several topical directories depending on their
purpose as part of the package. For example, several project control
variables are set to establish the session global environment in the "config"
directory rather than the "R" directory.
}
\details{
If a direct file match to name is not found, it will be searched for using a
recursive [list.files] allowing for regex matches (e.g. ".R$"). Directories
are similarly sought out within the project. Reasonable feedback is provided.

This convenience function uses [usethis::edit_file] to open (or create if
`create_new` is TRUE) any given file in the project.
}
\note{
If the directory and file cannot be found, and `create_new` is true,
  the directory will be placed within the project directory.
}
