% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_ums.R
\name{optimal_ums}
\alias{optimal_ums}
\title{Get the optimal uncertainty mass spectrum parameters for data}
\usage{
optimal_ums(
  peaktable,
  max_correl = 0.75,
  correl_bin = 0.05,
  max_ph = 10,
  ph_bin = 1,
  max_freq = 10,
  freq_bin = 1,
  min_n_peaks = 3,
  cormethod = "pearson"
)
}
\arguments{
\item{peaktable}{list generated from `create_peak_table_ms1` or `create_peak_table_ms2`}

\item{max_correl}{numeric maximum acceptable correlation}

\item{correl_bin}{numeric sequence bin width from max_correl..0}

\item{max_ph}{numeric maximum acceptable peak height (\%)}

\item{ph_bin}{numeric sequence bin width from max_ph..0}

\item{max_freq}{numeric maximum acceptable observational frequency (\%)}

\item{freq_bin}{numeric sequence bin width from max_freq..0}

\item{min_n_peaks}{integer ideal minimum number of scans for mass spectrum}

\item{cormethod}{string indicating correlation function to use (see [cor()] for description)}
}
\value{
data.frame object containing optimized search parameters
}
\description{
Get the optimal uncertainty mass spectrum parameters for data
}
