% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_comparison.R
\name{overlap}
\alias{overlap}
\title{Calculate overlap ranges}
\usage{
overlap(x1, e1, x2, e2)
}
\arguments{
\item{x1, x2}{values containing mean values}

\item{e1, e2}{values containing respective error values}
}
\description{
Internal function: determines if two ranges (x1-e1 to x1+e1) and (x2-e2 to x2+e2) overlap (nonstatistical evaluation)
}
