% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_comparison.R
\name{pair_ums}
\alias{pair_ums}
\title{Pairwise data.frame of two uncertainty mass spectra}
\usage{
pair_ums(ums1, ums2, error = 5, minerror = 0.002)
}
\arguments{
\item{ums1}{uncertainty mass spectrum from `get_ums` function}

\item{ums2}{uncertainty mass spectrum from `get_ums` function}

\item{minerror}{the minimum mass error (in Da) of the instrument data}

\item{masserror}{the mass accuracy (in ppm) of the instrument data}
}
\description{
The function stacks two uncertainty mass spectra together based on binned m/z values
}
