% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_gather_json.R
\name{peak_gather_json}
\alias{peak_gather_json}
\title{Extract peak data and metadata}
\usage{
peak_gather_json(
  methodjson,
  mzml,
  compoundtable,
  zoom = c(1, 5),
  minerror = 0.002
)
}
\arguments{
\item{methodjson}{list of JSON generated from `parse_method_json` function}

\item{mzml}{list of msdata from `mzMLtoR` function}

\item{compoundtable}{data.frame containing compound identities [should be extractable from SQL later]}

\item{zoom}{numeric vector specifying the range around the precursor ion to include, from m/z - zoom[1] to m/z + zoom[2]}

\item{minerror}{numeric the minimum error (in Da) of the instrument}
}
\value{
list of peak objects
}
\description{
gathers metadata from methodjson and extracts the MS1 and MS2 data from the mzml
}
