% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_comparison.R
\name{plot_compare_ms}
\alias{plot_compare_ms}
\title{Plot MS Comparison}
\usage{
plot_compare_ms(
  ums1,
  ums2,
  main = "Comparison Mass Spectrum",
  size = 1,
  c1 = "black",
  c2 = "red",
  ylim.exp = 1
)
}
\arguments{
\item{ums1, ums2}{uncertainty mass spectrum from `get_ums` function}

\item{main}{Main Title of the Plot}

\item{size}{line width of the mass spectra lines}

\item{c1}{Color of the top (ums1) mass spectral lines}

\item{c2}{Color of the bottom (ums2) mass spectral lines}

\item{ylim.exp}{Expansion unit for the y-axis}
}
\description{
Plots a butterfly plot for the comparison of two uncertainty mass spectra
}
