% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{plot_ms}
\alias{plot_ms}
\title{Generate consensus mass spectrum}
\usage{
plot_ms(
  ms,
  xlim = NULL,
  ylim = NULL,
  main = "Mass Spectrum",
  color = "black",
  size = 1,
  removal = 0
)
}
\arguments{
\item{peaklist}{result of the `create_peak_list` function}
}
\value{
nested list of dataframes containing all MS1 and MS2 data for the peak
}
\description{
TODO update documentation
}
\details{
The function extracts the relevant information and sorts it into nested lists for
use in the uncertainty functions
}
