% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_comparison.R
\name{pool.sd}
\alias{pool.sd}
\title{Pool standard deviations}
\usage{
pool.sd(sd, n)
}
\arguments{
\item{sd}{A vector containing numeric values of standard deviations}

\item{n}{A vector containing integers for the number of observations respective to the sd values}
}
\description{
Internal function: calculates a pooled standard deviation
}
