% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_comm.R
\name{pragma_table_def}
\alias{pragma_table_def}
\title{Get table definition from SQLite}
\usage{
pragma_table_def(db_table, db_conn = con, get_sql = FALSE, pretty = TRUE)
}
\arguments{
\item{db_table}{CHR vector name of the table(s) to inspect}

\item{db_conn}{connection object (default: con)}

\item{get_sql}{BOOL scalar of whether or not to return the schema sql
(default FALSE)}

\item{pretty}{BOOL scalar for whether to return "pretty" SQL that includes
human readability enhancements; if this is set to TRUE (the default), it is
recommended that the output is fed through `cat` and, in the case of
multiple tables}
}
\value{
data.frame object representing the SQL PRAGMA expression
}
\description{
Given a database connection (`con`). Get more information about the
properties of (a) database table(s) directly from `PRAGMA table_info()`
rather than e.g. [DBI::dbListFields()]. Set `get_sql` to `TRUE` to include
the direct schema using sqlite_master; depending on formatting this may or
may not be directly usable though some effort has been made to remove
formatting characters (e.g. line feeds, tabs, etc) if stringr is available.
}
\details{
Note that the package `stringr` is required for formatting returns that
include either `get_sql` or `pretty` as TRUE.
}
