% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_setup.R
\name{py_modules_available}
\alias{py_modules_available}
\title{Are all conda modules available in the active environment}
\usage{
py_modules_available("rdkit")
}
\arguments{
\item{required_modules}{CHR vector of required modules}
}
\value{
LGL scalar of whether or not all modules are available. Check console
  for further details.
}
\description{
Checks that all defined modules are available in the currently active python
binding. Supports error logging
}
