% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{qc_result}
\alias{qc_result}
\title{Does an import object have a defined QC value present?}
\usage{
qc_result(
  obj,
  qc_data_in = "qc",
  qc_method_in = "qcmethod",
  search_text_in = "name",
  search_text = "QC Method Used",
  value_in = "value"
)
}
\arguments{
\item{obj}{LIST import object of a single import file}

\item{qc_data_in}{CHR scalar name of an item in `obj` containing QC information
(default: "qc")}

\item{qc_method_in}{CHR scalar name of an item in `obj` containing QC method
results as a data.frame object of the form `data.frame(name = character(0),
value = logical(0))` (default: "qcmethod")}

\item{search_text_in}{CHR scalar column name in `obj[[qc_method_in]]`
containing the QC result names to look for (default: "name")}

\item{search_text}{CHR scalar of the QC result to look for (default: "QC
Method Used")}

\item{value_in}{CHR scalar column name in `obj[[qc_method_in]]` containing
the values of QC results in boolean format (default: "name")}
}
\value{
LGL scalar of the value, if present, as numeric, or 0 if not present
}
\description{
This returns a numerical boolean of whether or not the import has a given
structure for QC data, indicating whether the analytical method has a QC
method attached.
}
\examples{
qc_result(list(qc = list(letters), qcmethod = data.frame(name = "QC Method Used", value = TRUE)))

}
