% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_setup.R
\name{rdkit_mol_aliases}
\alias{rdkit_mol_aliases}
\title{Create aliases for a molecule from RDKit}
\usage{
rdkit_mol_aliases(
  identifiers,
  type = "smiles",
  mol_from_prefix = "MolFrom",
  get_aliases = c("inchi", "inchikey"),
  mol_to_prefix = "MolTo",
  rdkit_ref = "rdk",
  log_ns = "rdk",
  make_if_not = TRUE
)
}
\arguments{
\item{identifiers}{CHR vector of machine-readable molecule identifiers in a
format matching `type`}

\item{type}{CHR scalar of the type of encoding to use for `identifiers`
(default: smiles)}

\item{mol_from_prefix}{CHR scalar of the prefix to identify an RDKit function
to create a mol object from`identifiers` (default: "MolFrom")}

\item{get_aliases}{CHR vector of aliases to produce (default: c("inchi",
"inchikey"))}

\item{mol_to_prefix}{CHR scalar of the prefix to identify an RDKit function
to create an alias from a mol object (default: "MolTo")}

\item{rdkit_ref}{CHR scalar OR R object of an RDKit binding (default NULL
goes to "rdk" for convenience with other pipelines in this project)}

\item{log_ns}{}

\item{make_if_not}{LGL scalar of whether or not to create a new python
environment using [activate_py_env] if the binding is not active}
}
\value{
data.frame object containing the aliases and the original identifiers
}
\description{
Call this function to generate any number of machine-readable aliases from an
identifier set. Given the `identifiers` and their `type`, RDKit will be
polled for conversion functions to create a mol object. That mol object is
then used to create machine-readable aliases in any number of supported
formats. See the [RDKit Documentation](http://rdkit.org/docs/index.html) for
options. The `type` argument is used to match against a "MolFromX" funtion,
while the `aliases` argument is used to match against a "MolToX" function.
}
\details{
At the time of authorship, RDK v2021.09.4 was in use, which contained the
following options findable by this function: CMLBlock, CXSmarts, CXSmiles,
FASTA, HELM, Inchi, InchiAndAuxInfo, InchiKey, JSON, MolBlock, PDBBlock,
RandomSmilesVect, Sequence, Smarts, Smiles, TPLBlock, V3KMolBlock, XYZBlock.
}
\note{
Both `type` and `aliases` are case insensitive.

If `aliases` is set to NULL, all possible expressions (excluding those
  with "File" in the name) are returned from RDKit, which will likely produce
  NULL values and module ArgumentErrors.
}
