% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_functions.R
\name{read_log}
\alias{read_log}
\title{Read a log from a log file}
\usage{
read_log("log.txt")
}
\arguments{
\item{file}{CHR scalar file path to a log file (default NULL is translated to
"log.txt")}

\item{last_n}{INT scalar of the last 'n' log entries to read.}

\item{as_object}{LGL scalar of whether to return the log as an R object or
just to print the log to the console.}
}
\value{
CHR vector of the requested log file entries if `as_object` is TRUE,
  or none with a console print if `as_object` is FALSE
}
\description{
By default if `file` does not exist (i.e. `file` is not a fully defined path)
this looks for log text files in the directory defined by `LOG_DIRECTORY` in
the session.
}
