% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{rebuild_help_htmls}
\alias{rebuild_help_htmls}
\title{Rebuild the help files as HTML with an index}
\usage{
rebuild_help_htmls()
}
\value{
None
}
\description{
Rebuild the help files as HTML with an index
}
