% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{rectify_null_from_env}
\alias{rectify_null_from_env}
\title{Rectify NULL values provided to functions}
\usage{
rectify_null_from_env(test, test, "test")
}
\arguments{
\item{parameter}{the object being evaluated}

\item{env_parameter}{the name or object of a value to use from the
environment if \code{parameter} is NULL}

\item{default}{the fallback value to use if \code{parameter} is NULL and
\code{env_parameter} does not exist}

\item{log_ns}{the namespace to use with [log_it] if available}
}
\value{
The requested value, either as-is, rectified from the environment, or 
  the default
}
\description{
To support redirection of sensible parameter reads from an environment,
either Global or System, functions in this package may include NULL as their
default value. This returns values in precedence of \code{parameter},
\code{env_parameter} and \code{default}.
}
\note{
\code{log_ns} is only applicable if logging is set up in this project (see
  project settings in env_glob.txt, env_R.R, and env_logger.R for details).

Both [base::.GlobalEnv] and [base::Sys.getenv] are checked, and can be
  provided as a character scalar or as an object reference
}
