% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_comm.R
\name{ref_table_from_map}
\alias{ref_table_from_map}
\title{Get the name of a linked normalization table}
\usage{
ref_table_from_map("table1", "fk_column1", er_map(con), "references")
}
\arguments{
\item{table_name}{CHR scalar name of the database table}

\item{table_column}{CHR scalar name of the foreign key table column}

\item{this_map}{LIST object containing the schema representation from
`er_map` (default: an object named "db_map" created as part of the package
spin up)}

\item{fk_refs_in}{CHR scalar name of the item in `this_map` containing the
SQL "REFERENCES" statements extracted from the schema}
}
\value{
CHR scalar name of the table to which a FK column is linked or an
  empty character string if no match is located (i.e. `table_column` is not a
  defined foreign key).
}
\description{
Extract the name of a normalization table from the database given a table and
column reference.
}
\note{
This requires an object of the same shape and properties as those
  resulting from [er_map] as `this_map`.
}
