% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_comm.R
\name{remove_db}
\alias{remove_db}
\title{Remove an existing database}
\usage{
remove_db("test.sqlite", archive = TRUE)
}
\arguments{
\item{db}{CHR scalar name of the database to build (default: session value
DB_NAME)}

\item{archive}{LGL scalar of whether to create an archive of the current
database (if it exists) matching the name supplied in argument `db`
(default: FALSE)}
}
\value{
None, check console for details
}
\description{
This is limited to only the current working directory and includes its
subdirectories. If you wish to retain a copy of the prior database, ensure
argument `archive = TRUE` (note the default is FALSE) to create a copy of the
requested database prior to rebuild; this is created in the same directory as
the found database and appends
}
