% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{remove_icon_from}
\alias{remove_icon_from}
\title{Remove the last icon attached to an HTML element}
\usage{
remove_icon_from(id)
}
\arguments{
\item{id}{CHR scalar of the HTML ID from which to remove the last icon}
}
\value{
CHR scalar suitable to execute with `shinyjs::runJS`
}
\description{
Remove the last icon attached to an HTML element
}
\examples{
append_icon_to("example", "r-project", "fa-3x")
remove_icon_from("example")
}
