% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{remove_sample}
\alias{remove_sample}
\title{Delete a sample}
\usage{
remove_sample(sample_ids, db_conn = con, log_ns = "db")
}
\arguments{
\item{sample_ids}{INT vector of IDs to remove from the samples table.}

\item{db_conn}{connection object (default: con)}

\item{log_ns}{CHR scalar of the logging namespace to use (default: "db")}
}
\value{
None, executes actions on the database
}
\description{
Removes a sample from the database and associated records in ms_methods,
conversion_software_settings, and conversion_software_linkage. Associated
peak and mass spectrometric signals will also be removed.
}
