% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suspectlist.R
\name{repair_xl_casrn_forced_to_date}
\alias{repair_xl_casrn_forced_to_date}
\title{Repair CAS RNs forced to a date numeric by MSXL}
\usage{
repair_xl_casrn_forced_to_date(casrn_vec, output_format = "\%Y-\%m-\%d")
}
\arguments{
\item{casrn_vec}{CHR or NUM vector of what should be valid CAS RNs}

\item{output_format}{CHR scalar of the output format, which}
}
\value{
CHR vector of length equal to that of `casrn_vec` where numeric
  entries have been coerced to the assumed date
}
\description{
If a file is opened in Microsoft Excel(R), Chemical Abstract Service (CAS)
Registry Numbers (RNs) can occasionally be read as a pseudodate (e.g.
"1903-02-8"). Without tight controls over column formatting, this can result
in CAS RNs that are not real entering a processing pipeline. This convenience
function attempts to undo that automatic formatting by forcing vector members
whose values when coerced to numeric are equal to those provided to a
properly formatted date with an origin depending on operating system platform
(as read by `.Platform$OS.type`); Windows operating systems use the Windows
MSXL origin date of "1899-12-30" while others use "1904-01-01". Text entries
of "NA" are coerced to NA.
}
\examples{
repair_xl_casrn_forced_to_date(c("64324-08-3", "12332"))
}
