% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_gather_json.R
\name{repl_nan}
\alias{repl_nan}
\title{Replace NaN}
\usage{
repl_nan(x, repl = NULL)
}
\arguments{
\item{x}{vector of values}

\item{repl}{value to replace NaN contained in `x`}
}
\value{
vector with all NaN replaced with `repl`
}
\description{
Replace all NaN values with a specified value
}
