% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{reset_logger_settings}
\alias{reset_logger_settings}
\title{Update logger settings}
\usage{
reset_logger_settings()
}
\arguments{
\item{reload}{LGL scalar indicating (if TRUE) whether or not to refresh from
\code{env_R.R} or (if FALSE) to use the current environment settings (e.g. for
testing purposes) (default: FALSE)}
}
\value{
None
}
\description{
This is a simple action wrapper to update any settings that may have been
changed with regard to logger. If, for instance, something is not logging the
way you expect it to, change the relevant setting and then run
\code{update_logger_settings()} to reflect the current environment.
}
