% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_compound_fragments}
\alias{resolve_compound_fragments}
\title{Link together peaks, fragments, and compounds}
\usage{
resolve_compound_fragments(
  values = NULL,
  peak_id = NA,
  annotated_fragment_id = NA,
  compound_id = NA,
  linkage_table = "compound_fragments",
  peaks_table = "peaks",
  annotated_fragments_table = "annotated_fragments",
  compounds_table = "compounds",
  db_conn = con,
  log_ns = "db"
)
}
\arguments{
\item{values}{LIST item containing items for `peak_id`,
`annotated_fragment_id`, and `compound_id` (default: NULL); used
preferentially if provided}

\item{peak_id}{INT vector (ideally of length 1) of the peak ID(s) to link;
ignored if `values` is provided (default: NA)}

\item{annotated_fragment_id}{INT vector of fragment ID(s) to link; ignored if
`values` is provided (default: NA)}

\item{compound_id}{INT vector of compound ID(s) to link; ignored if `values`
is provided (default: NA)}

\item{linkage_table}{CHR scalar name of the database table containing
linkages between peaks, fragments, and compounds (default:
"compound_fragments")}

\item{peaks_table}{CHR scalar name of the database table containing peaks for
look up (default: "peaks")}

\item{compounds_table}{CHR scalar name of the table holding compound
information}

\item{db_conn}{connection object (default: con)}

\item{log_ns}{CHR scalar of the logging namespace to use (default: "db")}

\item{fragments_table}{CHR scalar name of the table holding annotated
fragment information}
}
\value{
None, value checks entries and executes database actions
}
\description{
This function links together the peaks, annotated_fragments, and compounds
table. This serves as the main connection table conceptually tying together
peaks, the fragments annotated within those peaks, and the compound
identification associated with the peaks. The database supports flexible
assignment wherein compounds may be related to either peaks or annotated
fragments, or both, and vice versa. At least two IDs are required for
linkage; i.e. compounds may not have an acciated peak in the database, but
are known to produce fragments at a particular m/z value. Ideally, all three
are provided to provide traceback from compounds, a complete list of their
annotated fragments, and association with a peak object with data containing
unannotated fragments, which can be traced back to the sample from which it
was drawn and the associated metrological method information.
}
