% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_compounds}
\alias{resolve_compounds}
\title{Resolve the compounds node during bulk import}
\usage{
resolve_compounds(
  obj,
  compounds_in = "compounddata",
  compounds_table = "compounds",
  compound_category = NULL,
  compound_category_table = "compound_categories",
  compound_alias_table = "compound_aliases",
  norm_alias_table = "norm_analyte_alias_references",
  norm_alias_name_column = "name",
  NIST_id_in = "id",
  require_all = FALSE,
  import_map = IMPORT_MAP,
  ensure_unique = TRUE,
  db_conn = con,
  log_ns = "db"
)
}
\arguments{
\item{obj}{LIST object containing data formatted from the import generator}

\item{compounds_in}{CHR scalar name in `obj` holding compound data (default:
"compounddata")}

\item{compounds_table}{CHR scalar name the database table holding compound
data (default: "compounds")}

\item{compound_category}{CHR or INT scalar of the compound category (either a
direct ID or a matching category label in `compound_category_table`)
(default: NULL)}

\item{compound_category_table}{CHR scalar name the database table holding
normalized compound categories (default: "compound_categories")}

\item{norm_alias_table}{CHR scalar name of the table normalizing analyte
alias references (default: "norm_analyte_alias_references")}

\item{norm_alias_name_column}{CHR scalar name of the column in
`norm_alias_table` containing the human-readable expression of alias type
classes (default: "name")}

\item{require_all}{LGL scalar of whether to require all columns (except the
assumed primary key column of "id") or only those defined as "NOT NULL"
(default: TRUE requires the presence of all columns in the table)}

\item{import_map}{data.frame object of the import map (e.g. from a CSV)}

\item{ensure_unique}{LGL scalar of whether or not to first check that the
values provided form a new unique record (default: TRUE)}

\item{db_conn}{connection object (default: con)}

\item{log_ns}{CHR scalar of the logging namespace to use (default: "db")}
}
\value{
INT scalar if successful, result of the call to [add_or_get_id]
  otherwise
}
\description{
Call this function as part of an import routine to resolve the compounds node.
}
\note{
This function is called as part of [full_import()]
}
