% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_description_NTAMRT}
\alias{resolve_description_NTAMRT}
\title{Resolve the method description tables during import}
\usage{
resolve_description_NTAMRT(
  obj,
  method_id,
  type = c("massspec", "chromatography"),
  mass_spec_in = "massspectrometry",
  chrom_spec_in = "chromatography",
  db_conn = con,
  fuzzy = TRUE,
  log_ns = "db"
)
}
\arguments{
\item{obj}{LIST object containing data formatted from the import generator}

\item{method_id}{INT scalar of the ms_method.id record to associate}

\item{type}{CHR scalar, one of "massspec" or "chromatography" depending on
the type of description to add; much of the logic is shared, only details
differ}

\item{mass_spec_in}{CHR scalar name of the element in `obj` holding mass
spectrometry information (default: "massspectrometry")}

\item{chrom_spec_in}{CHR scalar name of the element in `obj` holding
chromatographic information (default: "chromatography")}

\item{log_ns}{CHR scalar of the logging namespace to use during execution
(default: "db")}
}
\value{
None, executes actions on the database
}
\description{
Two tables (and their associated normalization tables) exist in the database
to store additional information about mass spectrometric and chromatographic
methods. These tables are "ms_descriptions" and "chromatography_descriptions"
and cannot be easily mapped directly. This function serves to coerce values
supplied during import into that required by the database. Primarily, the
issue rests in the need to support multiple descriptions of analytical
instrumentation (e.g. multiple mass analyzer types, multiple vendors,
multiple separation columns, etc.). Tables targeted by this function are
"long" tables that may well have `n` records for each mass spectrometric
method.
}
\note{
This function is called as part of [full_import()]

This function is brittle; built specifically for the NIST NTA MRT
  import format. If using a different import format, customize to your needs
  using this function as a guide.
}
