% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_method}
\alias{resolve_method}
\title{Add an ms_method record via import}
\usage{
resolve_method(
  obj,
  method_in = "massspectrometry",
  ms_methods_table = "ms_methods",
  db_conn = con,
  ensure_unique = TRUE,
  log_ns = "db",
  qc_data_in = "qc",
  qc_data_table = "qc_data",
  qc_method_in = "qcmethod",
  qc_method_table = "qc_methods",
  search_text_in = "name",
  search_text = "QC Method Used",
  value_in = "value",
  require_all = TRUE,
  import_map = IMPORT_MAP,
  ...
)
}
\arguments{
\item{obj}{LIST object containing data formatted from the import generator}

\item{method_in}{CHR scalar name of the `obj` list containing method
information}

\item{ms_methods_table}{CHR scalar name of the database table containing method
information}

\item{db_conn}{connection object (default: con)}

\item{ensure_unique}{LGL scalar of whether or not to first check that the
values provided form a new unique record (default: TRUE)}

\item{log_ns}{CHR scalar of the logging namespace to use (default: "db")}

\item{qc_data_in}{CHR scalar name of an item in `obj` containing QC information
(default: "qc")}

\item{qc_data_table}{CHR scalar name of the database table holding QC data
(default: "qc_data")}

\item{qc_method_in}{CHR scalar name of an item in `obj` containing QC method
results as a data.frame object of the form `data.frame(name = character(0),
value = logical(0))` (default: "qcmethod")}

\item{qc_method_table}{CHR scalar of the database table name holding QC
method check information (default: "qc_methods")}

\item{search_text_in}{CHR scalar column name in `obj[[qc_method_in]]`
containing the QC result names to look for (default: "name")}

\item{search_text}{CHR scalar of the QC result to look for (default: "QC
Method Used")}

\item{value_in}{CHR scalar column name in `obj[[qc_method_in]]` containing
the values of QC results in boolean format (default: "name")}

\item{require_all}{LGL scalar of whether to require all columns (except the
assumed primary key column of "id") or only those defined as "NOT NULL"
(default: TRUE requires the presence of all columns in the table)}

\item{import_map}{data.frame object of the import map (e.g. from a CSV)}

\item{...}{Other named elements to be appended to "ms_methods" as necessary
for workflow resolution, can be used to pass defaults or additional values.}
}
\value{
INT scalar if successful, result of the call to [add_or_get_id]
  otherwise
}
\description{
Part of the data import routine. Adds a record to the "ms_methods" table with
the values provided in the JSON import template. Makes extensive uses of
[resolve_normalization_value] to parse foreign key relationships.
}
\note{
This function is called as part of [full_import]
}
