% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_mobile_phase_NTAMRT}
\alias{resolve_mobile_phase_NTAMRT}
\title{Resolve the mobile phase node}
\usage{
resolve_mobile_phase_NTAMRT(
  obj,
  method_id,
  sample_id,
  peak_id,
  carrier_mix_names = NULL,
  id_mix_by = "^mp*[0-9]+",
  ms_methods_table = "ms_methods",
  sample_table = "samples",
  peak_table = "peaks",
  db_conn = con,
  mix_collection_table = "carrier_mix_collections",
  mobile_phase_props = list(in_item = "chromatography", db_table = "mobile_phases", props
    = c(flow = "flow", flow_units = "flowunits", duration = "duration", duration_units =
    "durationunits")),
  carrier_props = list(db_table = "carrier_mixes", norm_by = "norm_carriers", alias_in =
    "carrier_aliases", props = c(id_by = "solvent", fraction_by = "fraction")),
  additive_props = list(db_table = "carrier_additives", norm_by = "norm_additives",
    alias_in = "additive_aliases", props = c(id_by = "add$", amount_by = "_amount",
    units_by = "_units")),
  exclude_values = c("none", "", NA),
  fuzzy = TRUE,
  clean_up = TRUE,
  log_ns = "db"
)
}
\arguments{
\item{obj}{LIST object containing data formatted from the import generator}

\item{method_id}{INT scalar of the method id (e.g. from the import workflow)}

\item{sample_id}{INT scalar of the sample id (e.g. from the import workflow)}

\item{peak_id}{INT scalar of the peak id (e.g. from the import workflow)}

\item{carrier_mix_names}{CHR vector (optional) of carrier mix collection
names to assign, the length of which should equal 1 or the length of
discrete carrier mixtures; the default, NULL, will automatically assign
names as a function of the method and sample id.}

\item{id_mix_by}{CHR scalar regex to identify the elements of `obj` to use
for the mobile phase node (default "^mp*[0-9]+") grouping of carrier mix
collections, this is the main piece of connectivity pulling together the
descriptions and should only be changed to match different import naming
schemes}

\item{ms_methods_table}{CHR scalar name of the methods table (default:
"ms_methods")}

\item{sample_table}{CHR scalar name of the samples table (default: "samples")}

\item{peak_table}{CHR scalar name of the peaks table (default: "peaks")}

\item{db_conn}{existing connection object (e.g. of class "SQLiteConnection")}

\item{mix_collection_table}{CHR scalar name of the mix collections table
(default: "carrier_mix_collections")}

\item{mobile_phase_props}{LIST object describing how to import the mobile
phase table containing: in_item: CHR scalar name of the `obj` name
containing chromatographic information (default: "chromatography");
db_table: CHR scalar name of the mobile phases table (default:
"mobile_phases"); props: named CHR vector of name mappings with names equal
to database columns in `mobile_phase_props$db_table` and values matching
regex to match names in `obj[[mobile_phase_props$in_item]]`}

\item{carrier_props}{LIST object describing how to import the mobile phase
table containing: db_table: CHR scalar name of the mobile phases table
(default: "mobile_phases"); norm_table: CHR scalar name of the table used
to normalize carriers (default: "norm_carriers"); alias_table: CHR scalar
name of the table containing carrier aliases to search (default:
"carrier_aliases"); props: named CHR vector of name mappings with names
equal to database columns in `carrier_props$db_table` and values matching
regex to match names in `obj[[mobile_phase_props$in_item]]`, and an extra
element named `id_by` containing regex used to match names in the import
object indicate a carrier (e.g. "solvent")}

\item{additive_props}{LIST object describing how to import the mobile phase
table containing: db_table: CHR scalar name of the mobile phases table
(default: "mobile_phases"); norm_table: CHR scalar name of the table used
to normalize carriers (default: "norm_additives"); alias_table: CHR scalar
name of the table containing carrier aliases to search (default:
"additive_aliases"); props: named CHR vector of name mappings with names
equal to database columns in `additive_props$db_table` and values matching
regex to match names in `obj[[mobile_phase_props$in_item]]`
`obj[[mobile_phase_props$in_item]][[mobile_phase_props$db_table]]`, and an
extra element named `id_by` containing regex used to match names in the
import object indicate an additive (e.g. names terminating in "add")}

\item{exclude_values}{CHR vector indicating which values to ignore in `obj`
(default: c("none", "", NA))}

\item{fuzzy}{LGL scalar of whether to do a "fuzzy" match in the sense that
values provided are wrapped in an SQL SQL LIKE clause bookended with wildcards; 
overrides the `case_sensitive` setting if TRUE (default: FALSE).}

\item{clean_up}{LGL scalar determining whether or not to clean up the
`mix_collection_table` by removing just-added records if there are errors
adding to `carrier_props$db_table` (default: TRUE)}

\item{log_ns}{CHR scalar of the logging namespace to use during execution
(default: "db")}
}
\value{
None, executes actions on the database
}
\description{
The database node containing chromatographic method information is able to
handle any number of descriptive aspects regarding chromatography. It houses
normalized and aliased data in a manner that maximizes flexibility, allowing
any number of carrier agents (e.g. gasses for GC, solvents for LC) to be
described in increasing detail. To accommodate that, the structure itself may
be unintuitive and may not map well as records may be heavily nested.
}
\details{
The mobile phase node contains one record in table "mobile_phases" for each
method id, sample id, and carrier mix collection id with its associated flow
rate, normalized flow units, duration, and normalized duration units. Each
carrier mix collection has a name and child tables containing: records for
each value normalized carrier component and its unit fraction (e.g. in
carrier_mixes: Helium 1 would indicate pure Helium as a carrier gas in GC
work; Water, 0.9; Methanol, 0.1 to indicate a solvent mixture of 10% methanol
in water), as well as value normalized carrier additives, their amount, and
the units for that amount (mostly for LC work; e.g. in carrier_additives:
ammonium acetate, 5, mMol to indicate an additive to a solvent of 5 mMol
ammonium acetate); these are linked through the carrier mix collection id.

Call this function to import the results of the NIST Non-Targeted Analysis
Method Reporting Tool (NTA MRT), or feed it as `obj` a flat list containing
chromatography information.
}
\note{
This is a brittle function, and should only be used as part of the NTA
  MRT import process, or as a template for how to import data.

Some arguments are complicated by design to keep conceptual information
  together. These should be fed a structured list matching expectations. This
  applies to `mobile_phase_props`, `carrier_props`, and `additive_props`. See
  defaults in documentation for examples.

Database insertions are done in real time, so failures may result in
  hanging or orphaned records. Turn on `clean_up` to roll back by removing
  entries from `mix_collection_table` and relying on delete cascades built
  into the database. Additional names are provided here to match the schema.

This function is called as part of [full_import()]
}
