% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_ms_data}
\alias{resolve_ms_data}
\title{Resolve and store mass spectral data during import}
\usage{
resolve_ms_data(
  obj,
  peak_id = NULL,
  peaks_table = "peaks",
  ms_data_in = "msdata",
  ms_data_table = "ms_data",
  unpack_spectra = FALSE,
  ms_spectra_table = "ms_spectra",
  unpack_format = c("separated", "zipped"),
  as_object = FALSE,
  import_map = IMPORT_MAP,
  db_conn = con,
  log_ns = "db"
)
}
\arguments{
\item{obj}{LIST object containing data formatted from the import generator}

\item{peak_id}{INT scalar of the peak ID in question, which must be present}

\item{peaks_table}{CHR scalar name of the peaks table in the database
(default: "peaks")}

\item{ms_data_in}{CHR scalar of the named component of `obj` holding mass
spectral data (default: "msdata")}

\item{ms_data_table}{CHR scalar name of the table holding packed spectra in
the database (default: "ms_data")}

\item{unpack_spectra}{LGL scalar indicating whether or not to unpack spectral
data to a long format (i.e. all masses and intensities will become a single
record) in the table defined by `ms_spectra_table` (default: FALSE)}

\item{ms_spectra_table}{CHR scalar name of the table holding long form
spectra in the database (default: "ms_spectra")}

\item{unpack_format}{CHR scalar of the type of data packing for the spectra,
one of "separated" (default) or "zipped"}

\item{as_object}{LGL scalar indicating whether or not to return the result to
the session as an object (TRUE) or to add it to the database (default:
FALSE)}

\item{import_map}{data.frame object of the import map (e.g. from a CSV)}

\item{db_conn}{connection object (default: con)}

\item{log_ns}{CHR scalar of the logging namespace to use (default: "db")}
}
\value{
If `as_object` == TRUE, a data.frame object containing either packed
  (if `unpack_spectra` == FALSE) or unpacked (if `unpack_spectra` == TRUE)
  spectra, otherwise adds spectra to the database
}
\description{
Use peak IDs generated by the import workflow to assign and store mass
spectral data (if coming from the NIST NTA Method Reporting Tool, these will
all be in the "separated" format). Optionally also calls [resolve_ms_spectra]
if unpack_spectra = TRUE. Mass spectral data are stored in either one
("zipped")
}
\note{
This function is called as part of [full_import()] during the call to
  [resolve_peaks]
}
