% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_peak_ums_params}
\alias{resolve_peak_ums_params}
\title{Resolve and import optimal uncertain mass spectrum parameters}
\usage{
resolve_peak_ums_params(
  obj,
  peak_id,
  ums_params_in = "opt_ums_params",
  ums_params_table = "opt_ums_params",
  db_conn = con,
  log_ns = "db"
)
}
\arguments{
\item{obj}{LIST object containing data formatted from the import generator}

\item{peak_id}{INT scalar of the peak ID in question, which must be present
(e.g. from the import workflow)}

\item{ums_params_in}{CHR scalar name of the item in `obj` containing
optimized uncertainty mass spectrum parameters}

\item{ums_params_table}{CHR scalar name of the database table holding
optimized uncertainty mass spectrum parameters}

\item{log_ns}{CHR scalar of the logging namespace to use during execution
(default: "db")}
}
\value{
Nothing if successful, a data frame object of the extracted
  parameters otherwise.
}
\description{
This imports the defined object component containing parameters for the
optimized uncertainty mass spectrum used to compare with new data. This
function may be called at any time to add data for a given peak, but there is
no row unique restriction on the underlying table and is best used in a "one
pass" method during the import routine. These parameters are calculated as
part of NIST QA procedures and are added to the output of the NTA MRT after
those JSONs have been created.
}
\note{
This function is called as part of [resolve_peaks()]
}
