% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_peaks}
\alias{resolve_peaks}
\title{Resolve the peaks node during import}
\usage{
resolve_peaks(
  obj,
  sample_id,
  peaks_table = "peaks",
  software_timestamp = NULL,
  software_settings_in = "msconvertsettings",
  ms_data_in = "msdata",
  ms_data_table = "ms_data",
  unpack_spectra = FALSE,
  unpack_format = c("separated", "zipped"),
  ms_spectra_table = "ms_spectra",
  linkage_table = "conversion_software_peaks_linkage",
  settings_table = "conversion_software_settings",
  as_date_format = "\%Y-\%m-\%d \%H:\%M:\%S",
  format_checks = c("ymd_HMS", "ydm_HMS", "mdy_HMS", "dmy_HMS"),
  min_datetime = "2000-01-01 00:00:00",
  import_map = IMPORT_MAP,
  ums_params_in = "opt_ums_params",
  ums_params_table = "opt_ums_params",
  db_conn = con,
  log_ns = "db"
)
}
\arguments{
\item{obj}{CHR vector describing settings or a named LIST with names matching
column names in table conversion_software_settings.}

\item{sample_id}{INT scalar of the sample id (e.g. from the import workflow)}

\item{peaks_table}{CHR scalar of the database table name holding QC method check
information (default: "peaks")}

\item{ms_data_in}{CHR scalar of the named component of `obj` holding mass
spectral data (default: "msdata")}

\item{ms_data_table}{CHR scalar name of the table holding packed spectra in
the database (default: "ms_data")}

\item{unpack_spectra}{LGL scalar indicating whether or not to unpack spectral
data to a long format (i.e. all masses and intensities will become a single
record) in the table defined by `ms_spectra_table` (default: FALSE)}

\item{unpack_format}{CHR scalar of the type of data packing for the spectra,
one of "separated" (default) or "zipped"}

\item{ms_spectra_table}{CHR scalar name of the table holding long form
spectra in the database (default: "ms_spectra")}

\item{import_map}{data.frame object of the import map (e.g. from a CSV)}

\item{ums_params_in}{CHR scalar name of the item in `obj` containing
optimized uncertainty mass spectrum parameters}

\item{ums_params_table}{CHR scalar name of the database table holding
optimized uncertainty mass spectrum parameters}

\item{db_conn}{Connection object (default: con)}

\item{log_ns}{CHR scalar of the logging namespace to use (default: "db")}
}
\value{
INT scalar of the newly inserted or identified peak ID(s)
}
\description{
Call this function to resolve and insert information for the "peaks" node in
the database including software conversion settings (via
[resolve_software_settings_NTAMRT]) and mass spectra data (via
[resolve_ms_data] and, optionally, [resolve_ms_spectra]). This function
relies on the import object being formatted appropriately.
}
\note{
This function is called as part of [full_import()]

This function relies on an import map
}
