% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_qc_data_NTAMRT}
\alias{resolve_qc_data_NTAMRT}
\title{Resolve and import quality control data for import}
\usage{
resolve_qc_data_NTAMRT(
  obj,
  peak_id,
  qc_data_in = "qc",
  qc_data_table = "qc_data",
  peaks_table = "peaks",
  ignore = FALSE,
  db_conn = con,
  log_ns = "db"
)
}
\arguments{
\item{obj}{LIST object containing data formatted from the import generator}

\item{peak_id}{INT vector of the peak ids (e.g. from the import workflow)}

\item{qc_data_in}{CHR scalar name of the component in `obj` containing QC
data (default: "qc")}

\item{qc_data_table}{CHR scalar name of the database table holding QC data
(default: "qc_data")}

\item{peaks_table}{CHR scalar name of the database table holding peaks data
(default: "peaks")}

\item{log_ns}{CHR scalar of the logging namespace to use during execution
(default: "db")}
}
\value{
None, executes actions on the database
}
\description{
This imports the defined object component containing QC data (i.e. a nested
list of multiple quality control checks) from the NIST Non-Targeted Analysis
Method Reporting Tool (NTA MRT).
}
\note{
This function is called as part of [full_import()]
}
