% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_qc_methods_NTAMRT}
\alias{resolve_qc_methods_NTAMRT}
\title{Resolve and import quality control method information}
\usage{
resolve_qc_methods_NTAMRT(
  obj,
  peak_id,
  qc_method_in = "qcmethod",
  qc_method_table = "qc_methods",
  qc_method_norm_table = "norm_qc_methods_name",
  qc_method_norm_reference = "norm_qc_methods_reference",
  qc_references_in = "source",
  peaks_table = "peaks",
  ignore = FALSE,
  db_conn = con,
  log_ns = "db"
)
}
\arguments{
\item{obj}{LIST object containing data formatted from the import generator}

\item{peak_id}{INT vector of the peak ids (e.g. from the import workflow)}

\item{qc_method_in}{CHR scalar of the name in `obj` that contains QC method
check information (default: "qcmethod")}

\item{qc_method_table}{CHR scalar of the database table name holding QC
method check information (default: "qc_methods")}

\item{qc_method_norm_table}{CHR scalar name of the database table normalizing
QC methods type (default: "norm_qc_methods_name")}

\item{qc_method_norm_reference}{CHR scalar name of the database table
normalizing QC methods reference type (default:
"norm_qc_methods_reference")}

\item{qc_references_in}{CHR scalar of the name in `obj[[qc_method_in]]` that
contains the reference or citation for the QC protocol (default: "source")}

\item{peaks_table}{CHR scalar name of the database table holding sample
information (default: "samples")}

\item{log_ns}{CHR scalar of the logging namespace to use during execution
(default: "db")}
}
\value{
None, executes actions on the database
}
\description{
This imports the defined object component containing QC method information
(i.e. a data frame of multiple quality control checks) from the NIST
Non-Targeted Analysis Method Reporting Tool (NTA MRT).
}
\note{
This function is called as part of [full_import()]
}
