% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_sample}
\alias{resolve_sample}
\title{Add a sample via import}
\usage{
resolve_sample(
  obj,
  db_conn = con,
  method_id = NULL,
  sample_in = "sample",
  sample_table = "samples",
  generation_type = NULL,
  generation_type_default = "empirical",
  generation_type_norm_table = "norm_generation_type",
  import_map = IMPORT_MAP,
  ensure_unique = TRUE,
  require_all = TRUE,
  fuzzy = FALSE,
  case_sensitive = TRUE,
  log_ns = "db",
  ...
)
}
\arguments{
\item{obj}{LIST object containing data formatted from the import generator}

\item{db_conn}{connection object (default: con)}

\item{method_id}{INT scalar of the associated ms_methods record id}

\item{sample_in}{CHR scalar of the import object name storing sample data
(default: "sample")}

\item{sample_table}{CHR scalar name of the database table holding sample
information (default: "samples")}

\item{generation_type}{CHR scalar of the type of data generated for this
sample (e.g. "empirical" or "in silico"). The default (NULL) will assign
based on `generation_type_default`; any other value will override the
default value and be checked against values in `geneation_type_norm_table`}

\item{generation_type_default}{CHR scalar naming the default data generation
type (default: "empirical")}

\item{generation_type_norm_table}{CHR scalar name of the database table
normalizing sample generation type (default: "empirical")}

\item{import_map}{data.frame object of the import map (e.g. from a CSV)}

\item{ensure_unique}{LGL scalar of whether or not to first check that the
values provided form a new unique record (default: TRUE)}

\item{require_all}{LGL scalar of whether to require all columns (except the
assumed primary key column of "id") or only those defined as "NOT NULL"
(default: TRUE requires the presence of all columns in the table)}

\item{fuzzy}{LGL scalar of whether to do a "fuzzy" match in the sense that
values provided are wrapped in an SQL "LIKE '%value%'" clause; overrides
the `case_sensitive` setting if TRUE (default: FALSE).}

\item{case_sensitive}{LGL scalar of whether to match on a case sensitive
basis (the default TRUE searches for values as-provided) or whether to
coerce value matches by upper, lower, sentence, and title case matches}

\item{log_ns}{CHR scalar of the logging namespace to use (default: "db")}

\item{...}{Other named elements to be appended to samples as necessary for
workflow resolution, can be used to pass defaults or additional values.}
}
\value{
INT scalar if successful, result of the call to [add_or_get_id]
  otherwise
}
\description{
Part of the data import routine. Adds a record to the "samples" table with
the values provided in the JSON import template. Uses [verify_sample_class]
and [verify_contributor] to parse foreign key relationships, [resolve_method]
to add a record to ms_methods to get the proper id, and
[resolve_software_settings_NTAMRT] to insert records into and get the proper
conversion software linkage id from tables "conversion_software_settings" and
"conversion_software_linkage" if appropriate.
}
\note{
This function is called as part of [full_import()]
}
