% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_sample_aliases}
\alias{resolve_sample_aliases}
\title{Resolve and import sample aliases}
\usage{
resolve_sample_aliases(
  sample_id,
  obj = NULL,
  aliases_in = NULL,
  values = NULL,
  db_table = "sample_aliases",
  db_conn = con,
  log_ns = "db"
)
}
\arguments{
\item{sample_id}{INT scalar of the sample id (e.g. from the import workflow)}

\item{obj}{(optional) LIST object containing data formatted from the import
generator (default: NULL)}

\item{aliases_in}{(optional) CHR scalar of the name in `obj` containing the
sample aliases in list format (default: NULL)}

\item{values}{(optional) LIST containing the sample aliases with names as the
alias name and values containing the reference (e.g. URI, link to a
containing repository, or reference to the owner or project from which a
sample is drawn) to that alias}

\item{db_table}{CHR scalar name of the database table containing sample
aliases (default: "sample_aliases")}

\item{log_ns}{CHR scalar of the logging namespace to use during execution
(default: "db")}
}
\value{
None, executes actions on the database
}
\description{
Call this function to attach sample aliases to a sample record in the
database. This can be done either through the import object with a name
reference or directly by assigning additional values.
}
\note{
This function is called as part of [full_import()]

One of `values` or both of `obj` and `aliases_in` must be provided to
  add new sample aliases.
}
