% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_software_settings_NTAMRT}
\alias{resolve_software_settings_NTAMRT}
\title{Import software settings}
\usage{
resolve_software_settings_NTAMRT(
  obj,
  software_timestamp = NULL,
  db_conn = con,
  software_settings_in = "msconvertsettings",
  settings_table = "conversion_software_settings",
  linkage_table = "conversion_software_peaks_linkage",
  as_date_format = "\%Y-\%m-\%d \%H:\%M:\%S",
  format_checks = c("ymd_HMS", "ydm_HMS", "mdy_HMS", "dmy_HMS"),
  min_datetime = "2000-01-01 00:00:00",
  log_ns = "db"
)
}
\arguments{
\item{obj}{CHR vector describing settings or a named LIST with names matching
column names in table conversion_software_settings.}

\item{software_timestamp}{CHR scalar of the sample timestamp (e.g.
sample$starttime) to use for linking software conversion settings with peak
data, with a call back to the originating sample. If NULL (the default),
the current system timestamp in UTC will be used from [lubridate::now()].}

\item{db_conn}{connection object (default: con)}

\item{software_settings_in}{CHR scalar name of the component in `obj`
containing software settings (default: "msconvertsettings")}

\item{settings_table}{CHR scalar name of the database table containing the
software settings used for an imported data file (default:
"conversion_software_settings")}

\item{linkage_table}{CHR scalar name of the database table containing the
linkage between peaks and their software settings (default:
"conversion_software_peaks_linkage")}

\item{as_date_format}{CHR scalar the format to use when storing timestamps
that matches database column expectations (default: "\%Y-\%m-\%d \%H:\%M:\%S")}

\item{format_checks}{CHR vector of the [lubridate::parse_date_time()] format
checks to execute in order of priority; these must match a lubridate
function of the same name (default: c("ymd_HMS", "ydm_HMS", "mdy_HMS",
"dmy_HMS"))}

\item{min_datetime}{CHR scalar of the minimum reasonable timestamp used as a
sanity check (default: "2000-01-01 00:00:00")}

\item{log_ns}{CHR scalar of the logging namespace to use (default: "db")}
}
\value{
NULL on errors, INT scalar of the inserted software linkage id if
  successful
}
\description{
Part of the standard import pipeline, adding rows to the
`conversion_software_settings` table with a given sample id. Some argument
names are shared with other import functions, specifically `obj` but are
formed differently to resolve the node complexity correctly.
}
\note{
This function is called as part of [full_import()]
}
