% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_comm.R
\name{resolve_table_name}
\alias{resolve_table_name}
\title{Check presence of a database table}
\usage{
resolve_table_name(db_table = "compounds", db_conn = "test_con")
}
\arguments{
\item{db_table}{CHR vector of table names to check}

\item{db_conn}{connection object (default: con)}

\item{log_ns}{CHR scalar of the namespace (if any) to use for logging
(default: "db")}
}
\value{
CHR vector of existing tables
}
\description{
This convenience function checks for the existence of one or more `db_table`
objects in a database.
}
