% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_all}
\alias{search_all}
\title{Search all mass spectra within database against unknown mass spectrum}
\usage{
search_all(
  con,
  searchms,
  normfn = "sum",
  cormethod = "pearson",
  optimized_params = TRUE
)
}
\arguments{
\item{con}{SQLite database connection}

\item{searchms}{object generated from `create_search_ms` function}
}
\value{
LIST of search results
}
\description{
Search all mass spectra within database against unknown mass spectrum
}
