% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_precursor}
\alias{search_precursor}
\title{Search the database for all compounds with matching precursor ion m/z values}
\usage{
search_precursor(
  con,
  searchms,
  normfn = "sum",
  cormethod = "pearson",
  optimized_params = TRUE
)
}
\arguments{
\item{con}{SQLite database connection}

\item{searchms}{object generated from `create_search_ms` function}
}
\value{
table of match statistics for the compound of interest
}
\description{
Search the database for all compounds with matching precursor ion m/z values
}
