% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_setup.R
\name{setup_rdkit}
\alias{setup_rdkit}
\title{Conveniently set up an RDKit python environment for use with R}
\usage{
setup_rdkit(env_name = "nist_hrms_db", required_libraries = c("reticulate", "rdkit"), env_ref = "rdk")
}
\arguments{
\item{env_name}{CHR scalar of the name of a python environment}

\item{env_ref}{CHR scalar of the name of an R expression bound to a python
library OR an R object reference by name to an existing object that should be
bound to RDKit (e.g. from [reticulate::import])}

\item{ns}{CHR scalar}
}
\value{
None, though calls to utility functions will give their own returns
}
\description{
Conveniently set up an RDKit python environment for use with R
}
