% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smilestoformula.R
\name{smilestoformula}
\alias{smilestoformula}
\title{Convert SMILES string to Formula and other information}
\usage{
smilestoformula(SMILES)
}
\arguments{
\item{SMILES}{vector of SMILES strings}
}
\value{
data frame
}
\description{
The function converts SMILES strings into a data frame containing the molecular
formula (FORMULA), fixed mass of the formula (FIXED MASS), and the net charge (NETCHARGE).
}
\examples{
smilestoformula(c("CCCC", "C(F)(F)F"))

smilestoformula("CCCC")
}
