% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_msp.R
\name{sql_to_msp}
\alias{sql_to_msp}
\title{Export SQL Database to a MSP NIST MS Format}
\usage{
sql_to_msp(
  con,
  optimized_params = TRUE,
  outputfile = paste0("DimSpecExport", Sys.Date(), ".msp"),
  cormethod = "pearson",
  normfn = "sum"
)
}
\arguments{
\item{con}{SQLite database connection}

\item{optimized_params}{Boolean TRUE indicates that the optimized parameters for uncertainty mass spectra will be used.}

\item{outputfile}{Text string file name and/or location to save MSP file format}

\item{cormethod}{Text string type of correlation function to use (DEFAULT = 'pearson')}

\item{normfn}{Text string type of normalization function to use (DEFAULT = 'sum')}
}
\value{
None, saves a *.msp file to the local file system.
}
\description{
Export SQL Database to a MSP NIST MS Format
}
