% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_sqlite_scripts.R
\name{sqlite_auto_trigger}
\alias{sqlite_auto_trigger}
\title{Create a basic SQL trigger for handling foreign key relationships}
\usage{
sqlite_auto_trigger(target_table = "test", fk_col = c("col1", "col2",
  "col3"), norm_table = c("norm_col1", "norm_col2", "norm_col3"), pk_col =
  "id", val_col = "value", action_occurs = "after", trigger_action =
  "insert", table_action = "update")
}
\arguments{
\item{target_table}{CHR scalar name of a table with a foreign key constraint.}

\item{fk_col}{CHR vector name(s) of the column(s) in `target_table` with
foreign key relationship(s) defined.}

\item{norm_table}{CHR vector name(s) of the table(s) containing the primary
key relationship(s).}

\item{pk_col}{CHR vector name(s) of the column(s) in `norm_table` containing
the primary key(s) side of the relationship(s).}

\item{val_col}{CHR vector name(s) of the column(s) in `norm_table` containing
values related to the primary key(s) of the relationship(s).}

\item{action_occurs}{CHR scalar on when to run the trigger, must be one of
`c("before", "after", "instead")` ("instead" should only be used if
`target_table` is a view - this restriction is not enforced).}

\item{trigger_action}{CHR scalar on what type of trigger this is (e.g. `when`
= "after" and `trigger_action` = "insert" -> "AFTER INSERT INTO") and must
be one of `c("insert", "update", "delete")`.}

\item{for_each}{CHR scalar for SQLite this must be only `row` - translated
into a "FOR EACH ROW" clause. Set to any given noun for other SQL engines
supporting other trigger transaction types (e.g. "FOR EACH STATEMENT"
triggers)}

\item{table_action}{CHR scalar on what type of action to run when the trigger
fires, must be one of `c("insert", "update", "delete")`.}

\item{filter_col}{CHR scalar of a filter column to override the final WHERE
clause in the trigger. This should almost always be left as the default "".}

\item{filter_val}{CHR scalar of a filter value to override the final WHERE
clause in the trigger. This should almost always be left as the default "".}

\item{or_ignore}{LGL scalar on whether to ignore insertions to normalization
tables if an error occurs (default: TRUE, which can under certain
conditions raise exceptions during execution of the trigger if more than a
single value column exists in the parent table)}

\item{addl_actions}{CHR vector of additional target actions to add to
`table_action` statements, appended to the end of the resulting "insert" or
"update" actions to `target_table`. If multiple tables are in use, use
positional matching in the vector (e.g. with three normalization tables,
and additional actions to only the second, use c("", "additional actions",
""))}
}
\value{
CHR scalar of class glue containing the SQL necessary to create a
  trigger. This is raw text; it is not escaped and should be further
  manipulated (e.g. via dbplyr::sql()) as your needs and database
  communication pipelines dictate.
}
\description{
This creates a simple trigger designed to streamline foreign key compliance
for SQLite databases. Resulting triggers will check during table insert or
update actions that have one or more foreign key relationships defined as
`target_table.fk_col = norm_table.pk_col`. It is primarily for use in
controlled vocabulary lists where a single id is tied to a single value in
the parent table, but more complicated relationships can be handled.
}
\details{
These are intended as native database backup support for when connections do
not change the default SQLite setting of PRAGMA foreign_keys = off.
Theoretically any trigger could be created, but should only be used with care
outside the intended purpose.

Triggers created by this function will check all new INSERT and UPDATE
statements by checking provided values against their parent table keys. If an
index match is found no action will be taken on the parent table. If no match
is found, it is assumed this is a new normalized value and it will be added
to the normalization table and the resulting new key will be replaced in the
target table column.
}
\note{
While this will work on any number of combinations, all triggers should
  be heavily inspected prior to use. The default case for this trigger is to
  set it for a single FK/PK relationship with a single normalization value.
  It will run on any number of normalized columns however trigger behavior
  may be unexpected for more complex relationships.

If `or_ignore` is set to TRUE, errors in adding to the parent table
  will be ignored silently, possibly causing NULL values to be inserted into
  the target table foreign key column. For this reason it is recommended that
  the `or_ignore` parameter only be set to true to expand parent table
  entries, but it will only supply a single value for the new normalization
  table. If additional columns in the parent table must be populated (e.g.
  the parent table has two required columns "value" and "acronym"), it is
  recommended to take care of those prior to any action that would activate
  these triggers.

Parameters are not checked against a schema (e.g. tables and columns
  exist, or that a relationships exists between tables). This function
  processes only text provided to it.

Define individual relationships between `fk_col`, `norm_table`,
  `pk_col`, and `val_col` as necessary. Lengths for these parameters should
  match in a 1:1:1:1 manner to fully describe the relationships. If the
  schema of all tables listed in `norm_table` are close matches, e.g. all
  have two columns "id" and "value" then `pk_col` and `val_col` will be
  reused when only a single value is provided for them. That is, provided
  three `norm_table`(s) and one `pk_col` and one `val_col`, the arguments for
  `pk_col` and `val_col` will apply to each `norm_table`.

The usage example is built on a hypothetical SQLite schema containing
  four tables, one of which ("test" - with columns "id", "col1", "col2", and
  "col3") defines foreign key relationships to the other three ("norm_col1",
  "norm_col2", and "norm_col3").
}
\seealso{
build_triggers
}
