% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_sqlite_scripts.R
\name{sqlite_auto_view}
\alias{sqlite_auto_view}
\title{Create a basic SQL view of a normalized table}
\usage{
sqlite_auto_view(table_pragma = pragma_table_info("contributors"),
  target_table = "contributors", relationships =
  get_fkpk_relationships(db_map = er_map(con), dictionary =
  data_dictionary(con)), drop_if_exists = FALSE)
}
\arguments{
\item{table_pragma}{data.frame object from [pragma_table_info] for a given
table name in the database}

\item{target_table}{CHR scalar name of the database table to build for, which
should be present in the relationship definition}

\item{relationships}{data.frame object describing the foreign key
relationships for `target_table`, which should generally be the result of a
call to [get_fkpk_relationships]}

\item{drop_if_exists}{LGL scalar indicating whether to include a "DROP VIEW"
prefix for the generated view statement; as this has an impact on schema,
no default is set}
}
\value{
CHR scalar of class glue containing the SQL necessary to create a
  "denormalized" view. This is raw text; it is not escaped and should be
  further manipulated (e.g. via dbplyr::sql()) as your needs and database
  communication pipelines dictate.
}
\description{
Many database viewers will allow links for normalization tables to get the
human-readable value of a normalized column. Instead it is often preferable
to build in views automatically that "denormalize" such tables for display or
use in an application. This function seeks to script the process of creating
those views. It examines the table definition from [pragma_table_info] and
will extract the primary/foreign key relationships to build a "denormalized"
view of the table using [get_fkpk_relationships] which requires a database
map created from [er_map] and data dictionary created from [data_dictionary].
}
\details{
TODO for v2: abstract the relationships call by looking for objects in the
current session.
}
\note{
No schema checking is performed by this function, but rather relies on
  definitions from other functions.

This example will run slowly if the database map [er_map] and
  dictionary [data_dictionary] haven't yet been called. If they exist in your
  session, use those as arguments to get_fkpk_relationships.
}
\seealso{
build_views

pragma_table_info

get_fkpk_relationships

er_map

data_dictionary
}
