% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_comm.R
\name{sqlite_parse_import}
\alias{sqlite_parse_import}
\title{Parse SQL import statements}
\usage{
if (file.exists("./config/data/elements.csv")) {
  sqlite_parse_import(".import --csv --skip 1 ./config/data/elements.csv elements")
}
}
\arguments{
\item{build_statements}{CHR vector of SQL build statements from an SQL file.}
}
\value{
LIST of parsed .import statements as full "INSERT" statements.
}
\description{
In the absence of the sqlite command line interface (CLI), the [build_db]
process needs a full set of SQL statements to execute directly rather than
CLI dot commands. This utility function parses formatted SQL statements
containing CLI ".import" commands to create SQL INSERT statements. This
function is primarily called in [create_fallback_build].
}
