% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{start_api}
\alias{start_api}
\title{Start the plumber interface from a clean environment}
\usage{
start_api()
}
\arguments{
\item{plumber_file}{CHR scalar name of the plumber definition file, which
should be in \code{src_dir} (default: NULL)}

\item{plumber_host}{CHR scalar of the host server address (default: NULL)}

\item{plumber_port}{INT scalar of the listening port on the host server
(default: NULL)}

\item{background}{LGL scalar of whether to launch the API in a background
process (default: FALSE)}

\item{src_dir}{CHR scalar file path to settings and functions enabling the
plumber API (default: here::here("inst", "plumber"))}

\item{log_ns}{CHR scalar name of the logging namespace to use for this
function (default: "api")}
}
\value{
None, launches the plumber instance
}
\description{
This convenience function launches the plumber instance if it was not set to
launch during the session setup. It is a thin wrapper with a more intuitive
name than [api_reload] and the default background setting turned off to test
the server in the current session.
}
\note{
This function is intended to pull from the environment variables
  identifying the plumber file, host, and port.
}
