% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{summarize_check_fragments}
\alias{summarize_check_fragments}
\title{Summarize results of check_fragments function}
\usage{
summarize_check_fragments(fragments_checked)
}
\arguments{
\item{fragments_checked}{output of `check_fragments` function}
}
\value{
table summary of check_fragments function
}
\description{
Summarize results of check_fragments function
}
