% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{support_info}
\alias{support_info}
\title{R session information for support needs}
\usage{
support_info()
}
\arguments{
\item{app_info}{BOOL scalar on whether to return this application's properties}
}
\value{
LIST of values
}
\description{
Several items of interest for this particular project including:
- DB_DATE, DB_VERSION, BUILD_FILE, LAST_DB_SCHEMA, LAST_MODIFIED, DEPENDS_ON,
  and EXCLUSIONS as defined in the project's ../config/env_R.R file.
}
