% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{tack_on}
\alias{tack_on}
\title{Append additional named elements to a list}
\usage{
tack_on(obj, ..., log_ns = "db")
}
\arguments{
\item{obj}{LIST of any length to be appended to}

\item{...}{Additional arguments passed to/from the ellipsis parameter of
calling functions. If named, names are preserved.}

\item{log_ns}{CHR scalar of the logging namespace to use (default: "db")}
}
\value{
LIST object of length equal to \code{obj} plus additional named arguments
}
\description{
This does nothing more than [base::append] ellipsis arguments to be added
directly to the end of an existing list object. This primarily supports
additional property assignment during the import process for future
development and refinement. Call this as part of any function with additional
arguments. This may result in failures or ignoring unrecognized named
parameters. If no additional arguments are passed \code{obj} is returned as
provided.
}
\note{
If duplicate names exists in \code{obj} and those provided as ellipsis
  arguments, those provided as part of the ellipsis will replace those in
  \code{obj}.
}
\examples{
tack_on(list(a = 1:3), b = letters, c = rnorm(10))
tack_on(list(a = 1:3))
}
