% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_comm.R
\name{tidy_comments}
\alias{tidy_comments}
\title{Tidy up table and field comments}
\usage{
tidy_comments(pragma_table_def("compounds", get_sql = TRUE))
}
\arguments{
\item{obj}{result of calling [pragma_table_def] with `get_sql` = TRUE}
}
\value{
LIST of length equal to `obj` containing extracted comments
}
\description{
Creates more human-readable outputs after extracting the raw SQL used to
build entities and parsing out the comments as identified with the /* ... */
multi-line comment flag pair. Single line comments are not extracted. The
first comment is assumed to be the table comment. See examples in the
`config/sql_nodes` directory.
}
