% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_spectra.R
\name{tidy_ms_spectra}
\alias{tidy_ms_spectra}
\title{Tidy Spectra}
\usage{
tidy_ms_spectra(df = packed_data)
}
\arguments{
\item{df}{data.frame object containing nested spectra in a column}
}
\value{
data.frame object containing tidy spectra
}
\description{
A convenience function to take outputs from [ms_spectra_separated] and
[ms_spectra_zipped] and return them as a tidy data frame by unpacking the
list column "spectra".
}
