% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzML_base.R
\name{unzip}
\alias{unzip}
\title{Unzip binary data into vector}
\usage{
unzip(x, type = "gzip")
}
\arguments{
\item{x}{String of binary data to convert}

\item{type}{type of compression (see `base::memDecompress`). Default is `gzip`}
}
\value{
vector containing data from converted binary data
}
\description{
Unzip binary data into vector
}
