% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suspectlist.R
\name{update_data_sources}
\alias{update_data_sources}
\title{Dump current database contents}
\usage{
update_data_sources(
  project,
  data_dir = file.path("config", "data"),
  create_backups = TRUE,
  dump_tables = TRUE,
  dump_sql = TRUE,
  db_conn = con,
  sqlite_cli = ifelse(exists("SQLITE_CLI"), SQLITE_CLI, NULL),
  db_name = ifelse(exists("DB_NAME"), DB_NAME, NULL)
)
}
\arguments{
\item{project}{CHR scalar of the directory containing project specific data
(required, no default)}

\item{data_dir}{CHR scalar of the directory containing project independent
data sources used for population (default: `file.path("config", "data")`)}

\item{create_backups}{LGL scalar indicating whether to create backups prior
to writing updated data files (default: TRUE)}

\item{dump_tables}{LGL scalar indicating whether to dump contents of database
tables as comma-separated-value files (default: TRUE)}

\item{dump_sql}{LGL scalar indicating whether to create an SQL dump file
containing both schema and data as a backup (default: TRUE)}

\item{db_conn}{connection object (default: con)}

\item{SQLITE_CLI}{CHR scalar system reference to your installation of the
sqlite command line interface}
}
\value{
None, copies database information to the local file system
}
\description{
Perform one or both of two main tasks for backing up the NTA database.
}
\details{
The main task is to update CSV files in the config/data directory with the
current contents of the database. This is done on a table by table basis and
results in flat files whose structures no longer interrelate except
numerically. Primarily this would be used to migrate database contents to
other systems or for further manipulation. Please specify a `project` that
project-specific information can be maintained.

Backups created with this function are placed in a "backups" subdirectory of
the directory defined by parameter `data_dir`. If `dump_sql = TRUE` SQL dump
files will be written to "backups/sqlite" with file names equal to the
current database name prefixed by date.
}
