% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_setup.R
\name{update_env_from_file}
\alias{update_env_from_file}
\title{Update a conda environment from a requirements file}
\usage{
update_env_from_file("nist_hrms_db")
}
\arguments{
\item{env_name}{CHR scalar of a python environment}

\item{requirements_file}{CHR scalar file path to a suitable requirements.txt
or environment.yml file}

\item{conda_alias}{CHR scalar of the command line interface alias for your
conda tools (default: NULL is translated first to the environment variable
CONDA_CLI and then to "conda")}
}
\value{
None, directly updates the referenced python environment
}
\description{
The `requirements_file` can be any formatted file that contains a definition
for python libraries to add to an environment (e.g. requirements.txt,
environment.yml, etc) that is understood by conda. Relative file paths are
fine, but the file will not be discovered (e.g. by `list.files`) so
specificity is always better.
}
\details{
This is a helper function, largely to support versions of reticulate prior to
the introduction of the environment argument in version 1.24+.
}
\note{
This requires conda CLI tools to be installed.

A default installation alias of "conda" is assumed.

Set global variable `CONDA_CLI` to your conda alias for better support.
}
