% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_functions.R
\name{update_logger_settings}
\alias{update_logger_settings}
\title{Update logger settings}
\usage{
update_logger_settings(log_all_warnings = FALSE, log_all_errors =
  FALSE)
}
\arguments{
\item{log_all_warnings}{LGL scalar indicating whether or not to log all
warnings (default: TRUE)}

\item{log_all_errors}{LGL scalar indicating whether or not to log all errors
(default: TRUE)}
}
\value{
None
}
\description{
This applies the internal routing and formatting for logger functions to the
current value of the LOGGING object. If LOGGING is changed (i.e. a logging
namespace is added or changed) this function should be run to update routing
and formatting to be in line with the current settings.
}
\note{
The calling stack for auto logging of warnings and errors does not work
with background processes. These settings call [logger::log_warnings()] and
[logger::log_errors()].

This function is used only for its side effects.
}
