% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{user_guide}
\alias{user_guide}
\title{Launch the User Guide for DIMSpec}
\usage{
user_guide(path = "_book/index.html")
}
\arguments{
\item{path}{}
}
\value{
None, opens a browser to the index page of the User Guide
}
\description{
Use this function to launch the bookdown version of the User Guide for the
NIST Database Infrastructure for Mass Spectrometry (DIMSpec) Tool Set
}
\note{
This works ONLY when DIMSpec is used as a project with the defined
  directory structure
}
