% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suspectlist.R
\name{validate_casrns}
\alias{validate_casrns}
\title{Validate a CAS RN}
\usage{
validate_casrns(casrn_vec, strip_bad_cas = TRUE)
}
\arguments{
\item{casrn_vec}{CHR vector of what CAS RNs to validate}

\item{strip_bad_cas}{LGL scalar of whether to strip out invalid CAS RNs
(default: TRUE)}
}
\value{
CHR vector of length equal to that of `casrn_vec`
}
\description{
Chemical Abstract Service (CAS) Registry Numbers (RNs) follow a standard
creation format. From
[https://www.cas.org/support/documentation/chemical-substances/faqs], a CAS
RN is a "numeric identifier that can contain up to 10 digits, divided by
hyphens into three parts. The right-most digit is a check digit used to
verify the validity and uniqueness of the entire number. For example, 58-08-2
is the CAS Registry Number for caffeine."
}
\details{
Provided CAS RNs in `casrn_vec` are validated for format and their checksum
digit. Those failing will be printed to the console by default, and users
have the option of stripping unverified entries from the return vector.

This only validates that a CAS RN is properly constructed; it does not
indicate that the registry number exists in the CAS Registry.

See [repair_xl_casrn_forced_to_date] as one possible pre-processing step.
}
\examples{
validate_casrns(c("64324-08-9", "64324-08-5", "12332"))
validate_casrns(c("64324-08-9", "64324-08-5", "12332"), strip_bad_cas = FALSE)
}
