% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_generator.R
\name{validate_column_names}
\alias{validate_column_names}
\title{Ensure database column presence}
\usage{
validate_column_names(con, "peaks", "id")
}
\arguments{
\item{db_conn}{connection object (e.g. of class "SQLiteConnection")}

\item{table_names}{CHR vector of tables to search}

\item{column_names}{CHR vector of column names to validate}
}
\value{
None
}
\description{
When working with SQL databases, this convenience function validates any
number of column names by comparing against the list of column names in any
number of tables. Typically it is called transparently inline to cause
execution failure when column names are not present in referenced tables
during build of SQL queries.
}
