% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_generator.R
\name{validate_tables}
\alias{validate_tables}
\title{Ensure database table presence}
\usage{
validate_tables(con, "peaks")
}
\arguments{
\item{db_conn}{connection object (e.g. of class "SQLiteConnection")}

\item{table_names}{CHR vector name of tables to ensure are present}
}
\value{
Failure if the table doesn't exist, none if it does.
}
\description{
When working with SQL databases, this convenience function validates any
number of table names by comparing against the list of those present.
Typically it is called transparently inline to cause execution failure when
tables are not present during build of SQL queries.
}
