% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{with_help}
\alias{with_help}
\title{Convenience application of `add_help` using pipes directly in UI.R}
\usage{
with_help(widget, tooltip, ...)
}
\arguments{
\item{widget}{shiny.tag widget}

\item{tooltip}{CHR scalar of the tooltip text}

\item{...}{Other named arguments to be passed to `add_help`}
}
\value{

}
\description{
This may not work for certain widgets with heavily nested HTML. Note that
classes may be CSS dependent.
}
\examples{
actionButton("example", "With Help") \%>\%
  with_help("Now with a question mark icon hosting a tooltip")
actionButton("example", "With Help") \%>\%
  with_help("Large and green", size = "xl", class = "success")

}
