% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_search_object}
\alias{get_search_object}
\title{Generate msdata object from input peak data}
\usage{
get_search_object(searchmzml, zoom = c(1, 4))
}
\arguments{
\item{searchmzml}{mzml with searching dataframe from `getmzML` function}

\item{zoom}{vector length of 2 containing +/- the area around the MS1 precursor ion to collect data.}
}
\value{
LIST object of data.frames include MS1 and MS2 analytical data, and the search parameters used to generate them
}
\description{
Generate msdata object from input peak data
}
