% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopic_distribution.R
\name{isotopic_distribution}
\alias{isotopic_distribution}
\title{Isotopic distribution functions
Generate isotopic distribution mass spectrum of elemental formula}
\usage{
isotopic_distribution(
  elementalformula,
  exactmasschart,
  remove.elements = c(),
  max.dist = 3,
  min.int = 0.001,
  charge = "neutral"
)
}
\arguments{
\item{elementalformula}{character string of elemental formula to simulate isotopic pattern}

\item{exactmasschart}{exact mass chart generated from function create_exactmasschart}

\item{remove.elements}{character vector of elements to remove from elemental formula}

\item{max.dist}{numeric maximum mass distance (in Da) from exact mass to include in simulated isotopic pattern}

\item{min.int}{numeric minimum relative intensity (maximum = 1, minimum = 0) to include in simulated isotopic pattern}

\item{charge}{character string for the charge state of the simulated isotopic pattern, options are `neutral`, `positive`, and `negative`}
}
\value{
data frame containing mz and int values of mass spectrum
}
\description{
Isotopic distribution functions
Generate isotopic distribution mass spectrum of elemental formula
}
