% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{rebuild_help_htmls}
\alias{rebuild_help_htmls}
\title{Rebuild the help files as HTML with an index}
\usage{
rebuild_help_htmls(rebuild_book = TRUE, book = "dimspec_user_guide")
}
\arguments{
\item{rebuild_book}{LGL scalar of whether or not to rebuild an associated bookdown document}

\item{book}{Path to folder containing the bookdown document to rebuild}
}
\value{
URL to the requested book
}
\description{
Rebuild the help files as HTML with an index
}
