% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST_import_routines.R
\name{resolve_method}
\alias{resolve_method}
\title{Add an ms_method record via import}
\usage{
resolve_method(
  obj,
  method_in = "massspectrometry",
  ms_methods_table = "ms_methods",
  db_conn = con,
  ensure_unique = TRUE,
  log_ns = "db",
  qc_method_in = "qcmethod",
  qc_search_text = "QC Method Used",
  qc_value_in = "value",
  require_all = TRUE,
  import_map = IMPORT_MAP,
  ...
)
}
\arguments{
\item{obj}{LIST object containing data formatted from the import generator}

\item{method_in}{CHR scalar name of the `obj` list containing method
information}

\item{ms_methods_table}{CHR scalar name of the database table containing
method information}

\item{db_conn}{connection object (default: con)}

\item{ensure_unique}{LGL scalar of whether or not to first check that the
values provided form a new unique record (default: TRUE)}

\item{log_ns}{CHR scalar of the logging namespace to use (default: "db")}

\item{qc_method_in}{CHR scalar name of the import object element containing
QC method information (default: "qcmethod")}

\item{qc_search_text}{CHR scalar name of an element in the import object in
part `qc_method_in` identifying whether or not a QC method was used
(default: "QC Method Used")}

\item{qc_value_in}{CHR scalar name of an element in the import object
corresponding to `qc_method_in` where the value of the metric named for
`qc_search_text` is located (default: "value")}

\item{require_all}{LGL scalar of whether to require all columns (except the
assumed primary key column of "id") or only those defined as "NOT NULL"
(default: TRUE requires the presence of all columns in the table)}

\item{import_map}{data.frame object of the import map (e.g. from a CSV)}

\item{...}{Other named elements to be appended to "ms_methods" as necessary
for workflow resolution, can be used to pass defaults or additional values.}
}
\value{
INT scalar if successful, result of the call to [add_or_get_id]
  otherwise
}
\description{
Part of the data import routine. Adds a record to the "ms_methods" table with
the values provided in the JSON import template. Makes extensive uses of
[resolve_normalization_value] to parse foreign key relationships.
}
\note{
This function is called as part of [full_import()]
}
