% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_all}
\alias{search_all}
\title{Search all mass spectra within database against unknown mass spectrum}
\usage{
search_all(
  con,
  searchms,
  normfn = "sum",
  cormethod = "pearson",
  optimized_params = TRUE
)
}
\arguments{
\item{con}{SQLite database connection}

\item{searchms}{object generated from `create_search_ms` function}

\item{normfn}{the normalization function typically "mean" or "sum" for normalizing the intensity values}

\item{cormethod}{the correlation method used for calculating the correlation, see `cor` function for methods}

\item{optimized_params}{LGL scalar indicating whether or not to use the
optimal search parameters stored in the database table `opt_ums_params`}
}
\value{
LIST of search results
}
\description{
Search all mass spectra within database against unknown mass spectrum
}
