% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{user_guide}
\alias{user_guide}
\title{Launch the User Guide for DIMSpec}
\usage{
user_guide()
}
\arguments{
\item{path}{CHR scalar representing a valid file path to the local user guide}

\item{url_gh}{CHR scalar pointing to the web resource, in this case the URL
to the User Guide hosted on GitHub pages}

\item{view_on_github}{LGL scalar of whether to use the hosted version of the
User Guide on GitHub (default: TRUE is recommended) which will always
display the most up to date version}
}
\value{
None, opens a browser to the index page of the User Guide
}
\description{
Use this function to launch the bookdown version of the User Guide for the
NIST Database Infrastructure for Mass Spectrometry (DIMSpec) Toolkit
}
\note{
This works ONLY when DIMSpec is used as a project with the defined
  directory structure
}
